Dose-response model function for the qraLm models

This function provides the marginal probability of invasive listeriosis in a given population for a given Dose in CFU using the JEMRA, the Pouillot, the Fritsch or the EFSA dose-response models or the model developed within this project (EFSAMV,EFSAV,EFSALV) (see References).

Usage

DRForModel(data = list(), model = "JEMRA", population = 1, Poisson = FALSE)

Arguments

data

a list of (a minima):

N

(CFU) A matrix containing the numbers of L. monocytogenes per portion of defrosted or frozen vegetables.

model

either JEMRA, Pouillot, Fritsch, EFSA, EFSAMV,EFSAV or EFSALV

population

considered population (scalar).

Poisson

if TRUE, assume that Dose is the mean of a Poisson distribution. (actual LogNormal Poisson). If FALSE (default), assume that Dose is the actual number of bacteria.

Value

the data object with added:

Risk

A matrix of risk, of size similar to N

Model

the Model

Population

the Population

Details

see doseresponsemodels::DR() or doseresponsemodels::DRQuick() for details on population and models.

Note

This function uses (for all model but JEMRA) a linear approximation (approxfun) from the exact doseresponsemodels::DR() model evaluated on \(Dose = c(0,10^{seq(-5,12,length=1701)})\) (if Poisson=TRUE) or \(c(0,10^{seq(0,12,length=2000)})\) (if Poisson=FALSE). Any Dose lower or higher than these ranges will lead to NA.

References

EFSA (2018). “Scientific opinion on the Listeria monocytogenes contamination of ready-to-eat foods and the risk from human health in the EU.” EFSA Journal, 16(1), 5134. FAO-WHO (2004). “Risk assessment of Listeria monocytogenes in ready-to-eat foods: Technical report.” World Health Organization and Food and Agriculture Organization of the United Nations. Fritsch L, Guillier L, Augustin J (2018). “Next generation quantitative microbiological risk assessment: Refinement of the cold smoked salmon-related listeriosis risk model by integrating genomic data.” Microbial Risk Analysis, 10, 20–27. doi:10.1016/j.mran.2018.06.003 . Pouillot R, Hoelzer K, Chen Y, Dennis SB (2015). “Listeria monocytogenes dose response revisited–incorporating adjustments for variability in strain virulence and host susceptibility.” Risk Analysis, 35(1), 90–108. doi:10.1111/risa.12235 .

Author

Regis Pouillot, Vasco Cadavez

Examples

data <- list(N = matrix(10^runif(100, 0, 5), ncol = 25))
DRForModel(data, "JEMRA", 1)
#> $N
#>           [,1]        [,2]         [,3]       [,4]        [,5]         [,6]
#> [1,] 53.598321 20022.62225     5.056512   1.136298  316.498590 85296.485139
#> [2,]  8.927479    98.91643 51976.233752 361.667636    6.069254  1450.337512
#> [3,] 22.016464 32711.48599     2.879875 451.689103 7625.325165  5578.258050
#> [4,] 24.602063    44.85786 10723.601382   9.281161 3218.419023     3.372205
#>             [,7]        [,8]        [,9]       [,10]      [,11]        [,12]
#> [1,] 1374.500784    5.620047 21321.30520   69.072473   185.0684 20151.564559
#> [2,]    2.951276    8.894905    27.02070   13.983881   227.9102   618.177242
#> [3,]    1.859273 1656.944299    36.15272    3.178876 17563.0403     4.532657
#> [4,]  568.273614 7893.207343 80182.39410 1577.923195   539.9458     1.800693
#>           [,13]        [,14]     [,15]    [,16]     [,17]      [,18]
#> [1,] 118.127731 23074.510728 176.61110 848.3881 18729.779  7237.8229
#> [2,]   4.996841   156.813354   3.07324 319.2091    68.288 31571.8822
#> [3,] 435.291317   574.240689  52.29830 103.0539  6260.446   398.7185
#> [4,]  86.834488     2.952929 490.63343 450.2086  2302.807  1974.2504
#>             [,19]       [,20]      [,21]        [,22]       [,23]       [,24]
#> [1,] 22799.583298  2647.92048 3812.81404    15.677384 12773.89322    3.326025
#> [2,]     3.960392   135.84231  178.00621 59542.678292  4606.12891 8783.174606
#> [3,]   118.631649 12460.42728 4768.38442    17.660218    72.03876  125.327443
#> [4,]    81.739275    32.31576   25.47436     5.653791 10404.50621 1015.489357
#>             [,25]
#> [1,]     4.941774
#> [2,]     1.568708
#> [3,]   947.443166
#> [4,] 39794.681507
#> 
#> $Risk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 1.267431e-12 4.734900e-10 1.195710e-13 2.686740e-14 7.484457e-12
#> [2,] 2.111644e-13 2.339129e-12 1.229121e-09 8.552603e-12 1.435518e-13
#> [3,] 5.206946e-13 7.735531e-10 6.805667e-14 1.068146e-11 1.803218e-10
#> [4,] 5.817569e-13 1.060818e-12 2.535890e-10 2.194911e-13 7.610834e-11
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 2.017070e-09 3.250389e-11 1.328937e-13 5.042009e-10 1.633360e-12
#> [2,] 3.429723e-11 6.983303e-14 2.103873e-13 6.389334e-13 3.307354e-13
#> [3,] 1.319133e-10 4.396483e-14 3.918299e-11 8.549828e-13 7.516210e-14
#> [4,] 7.971401e-14 1.343836e-11 1.866566e-10 1.896133e-09 3.731437e-11
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 4.376499e-12 4.765391e-10 2.793432e-12 5.456603e-10 4.176437e-12
#> [2,] 5.389578e-12 1.461853e-11 1.181277e-13 3.708256e-12 7.271961e-14
#> [3,] 4.153264e-10 1.071365e-13 1.029365e-11 1.357947e-11 1.236788e-12
#> [4,] 1.276845e-11 4.263256e-14 2.053469e-12 6.983303e-14 1.160239e-11
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 2.006251e-11 4.429171e-10 1.711582e-10 5.391588e-10 6.261736e-11
#> [2,] 7.548628e-12 1.614819e-12 7.466040e-10 9.370282e-14 3.212319e-12
#> [3,] 2.436940e-12 1.480455e-10 9.428791e-12 2.805423e-12 2.946611e-10
#> [4,] 1.064637e-11 5.445622e-11 4.668654e-11 1.932898e-12 7.641665e-13
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 9.016443e-11 3.707035e-13 3.020738e-10 7.860379e-14 1.169065e-13
#> [2,] 4.209411e-12 1.408050e-09 1.089245e-10 2.077023e-10 3.708145e-14
#> [3,] 1.127616e-10 4.176659e-13 1.703526e-12 2.963740e-12 2.240486e-11
#> [4,] 6.024070e-13 1.336709e-13 2.460432e-10 2.401401e-11 9.410547e-10
#> 
#> $lotMeanRisk
#> [1] 2.280077e-10 1.515128e-10 8.701552e-11 1.527221e-10
#> 
#> $servingRisk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 1.267431e-12 4.734900e-10 1.195710e-13 2.686740e-14 7.484457e-12
#> [2,] 2.111644e-13 2.339129e-12 1.229121e-09 8.552603e-12 1.435518e-13
#> [3,] 5.206946e-13 7.735531e-10 6.805667e-14 1.068146e-11 1.803218e-10
#> [4,] 5.817569e-13 1.060818e-12 2.535890e-10 2.194911e-13 7.610834e-11
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 2.017070e-09 3.250389e-11 1.328937e-13 5.042009e-10 1.633360e-12
#> [2,] 3.429723e-11 6.983303e-14 2.103873e-13 6.389334e-13 3.307354e-13
#> [3,] 1.319133e-10 4.396483e-14 3.918299e-11 8.549828e-13 7.516210e-14
#> [4,] 7.971401e-14 1.343836e-11 1.866566e-10 1.896133e-09 3.731437e-11
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 4.376499e-12 4.765391e-10 2.793432e-12 5.456603e-10 4.176437e-12
#> [2,] 5.389578e-12 1.461853e-11 1.181277e-13 3.708256e-12 7.271961e-14
#> [3,] 4.153264e-10 1.071365e-13 1.029365e-11 1.357947e-11 1.236788e-12
#> [4,] 1.276845e-11 4.263256e-14 2.053469e-12 6.983303e-14 1.160239e-11
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 2.006251e-11 4.429171e-10 1.711582e-10 5.391588e-10 6.261736e-11
#> [2,] 7.548628e-12 1.614819e-12 7.466040e-10 9.370282e-14 3.212319e-12
#> [3,] 2.436940e-12 1.480455e-10 9.428791e-12 2.805423e-12 2.946611e-10
#> [4,] 1.064637e-11 5.445622e-11 4.668654e-11 1.932898e-12 7.641665e-13
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 9.016443e-11 3.707035e-13 3.020738e-10 7.860379e-14 1.169065e-13
#> [2,] 4.209411e-12 1.408050e-09 1.089245e-10 2.077023e-10 3.708145e-14
#> [3,] 1.127616e-10 4.176659e-13 1.703526e-12 2.963740e-12 2.240486e-11
#> [4,] 6.024070e-13 1.336709e-13 2.460432e-10 2.401401e-11 9.410547e-10
#> 
#> $Model
#> [1] "JEMRA"
#> 
#> $Population
#> [1] 1
#> 
DRForModel(data, "EFSAMV", 10)
#> $N
#>           [,1]        [,2]         [,3]       [,4]        [,5]         [,6]
#> [1,] 53.598321 20022.62225     5.056512   1.136298  316.498590 85296.485139
#> [2,]  8.927479    98.91643 51976.233752 361.667636    6.069254  1450.337512
#> [3,] 22.016464 32711.48599     2.879875 451.689103 7625.325165  5578.258050
#> [4,] 24.602063    44.85786 10723.601382   9.281161 3218.419023     3.372205
#>             [,7]        [,8]        [,9]       [,10]      [,11]        [,12]
#> [1,] 1374.500784    5.620047 21321.30520   69.072473   185.0684 20151.564559
#> [2,]    2.951276    8.894905    27.02070   13.983881   227.9102   618.177242
#> [3,]    1.859273 1656.944299    36.15272    3.178876 17563.0403     4.532657
#> [4,]  568.273614 7893.207343 80182.39410 1577.923195   539.9458     1.800693
#>           [,13]        [,14]     [,15]    [,16]     [,17]      [,18]
#> [1,] 118.127731 23074.510728 176.61110 848.3881 18729.779  7237.8229
#> [2,]   4.996841   156.813354   3.07324 319.2091    68.288 31571.8822
#> [3,] 435.291317   574.240689  52.29830 103.0539  6260.446   398.7185
#> [4,]  86.834488     2.952929 490.63343 450.2086  2302.807  1974.2504
#>             [,19]       [,20]      [,21]        [,22]       [,23]       [,24]
#> [1,] 22799.583298  2647.92048 3812.81404    15.677384 12773.89322    3.326025
#> [2,]     3.960392   135.84231  178.00621 59542.678292  4606.12891 8783.174606
#> [3,]   118.631649 12460.42728 4768.38442    17.660218    72.03876  125.327443
#> [4,]    81.739275    32.31576   25.47436     5.653791 10404.50621 1015.489357
#>             [,25]
#> [1,]     4.941774
#> [2,]     1.568708
#> [3,]   947.443166
#> [4,] 39794.681507
#> 
#> $Risk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 1.295678e-09 4.840214e-07 1.222353e-10 2.746869e-11 7.650988e-09
#> [2,] 2.158115e-10 2.391191e-09 1.256449e-06 8.742898e-09 1.467172e-10
#> [3,] 5.322227e-10 7.907560e-07 6.961767e-11 1.091906e-08 1.843331e-07
#> [4,] 5.947266e-10 1.084387e-09 2.592300e-07 2.243614e-10 7.780151e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 2.061900e-06 3.322696e-08 1.358582e-10 5.154153e-07 1.669747e-09
#> [2,] 3.506023e-08 7.134371e-11 2.150241e-10 6.531943e-10 3.380442e-10
#> [3,] 1.348478e-07 4.494579e-11 4.005470e-08 8.739505e-10 7.684567e-11
#> [4,] 8.151917e-11 1.373736e-08 1.908088e-07 1.938278e-06 3.814446e-08
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 4.473816e-09 4.871384e-07 2.855602e-09 5.577965e-07 4.269370e-09
#> [2,] 5.509467e-09 1.494372e-08 1.207929e-10 3.790782e-09 7.429204e-11
#> [3,] 4.245644e-07 1.095718e-10 1.052267e-08 1.388161e-08 1.264251e-09
#> [4,] 1.305257e-08 4.352967e-11 2.099124e-09 7.138366e-11 1.186050e-08
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 2.050880e-08 4.527687e-07 1.749657e-07 5.511505e-07 6.401039e-08
#> [2,] 7.716513e-09 1.650784e-09 7.632079e-07 9.573791e-11 3.283831e-09
#> [3,] 2.491210e-09 1.513388e-07 9.638559e-09 2.867783e-09 3.012155e-07
#> [4,] 1.088327e-08 5.566767e-08 4.772521e-08 1.975953e-09 7.811964e-10
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 9.217030e-08 3.789827e-10 3.087932e-07 8.040282e-11 1.194617e-10
#> [2,] 4.303095e-09 1.439354e-06 1.113477e-07 2.123227e-07 3.792170e-11
#> [3,] 1.152701e-07 4.269154e-10 1.741454e-09 3.029647e-09 2.290334e-08
#> [4,] 6.158134e-10 1.366739e-10 2.515163e-07 2.454828e-08 9.619810e-07
#> 
#> $lotMeanRisk
#> [1] 2.330778e-07 1.548817e-07 8.895097e-08 1.561178e-07
#> 
#> $servingRisk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 1.295678e-09 4.840214e-07 1.222353e-10 2.746869e-11 7.650988e-09
#> [2,] 2.158115e-10 2.391191e-09 1.256449e-06 8.742898e-09 1.467172e-10
#> [3,] 5.322227e-10 7.907560e-07 6.961767e-11 1.091906e-08 1.843331e-07
#> [4,] 5.947266e-10 1.084387e-09 2.592300e-07 2.243614e-10 7.780151e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 2.061900e-06 3.322696e-08 1.358582e-10 5.154153e-07 1.669747e-09
#> [2,] 3.506023e-08 7.134371e-11 2.150241e-10 6.531943e-10 3.380442e-10
#> [3,] 1.348478e-07 4.494579e-11 4.005470e-08 8.739505e-10 7.684567e-11
#> [4,] 8.151917e-11 1.373736e-08 1.908088e-07 1.938278e-06 3.814446e-08
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 4.473816e-09 4.871384e-07 2.855602e-09 5.577965e-07 4.269370e-09
#> [2,] 5.509467e-09 1.494372e-08 1.207929e-10 3.790782e-09 7.429204e-11
#> [3,] 4.245644e-07 1.095718e-10 1.052267e-08 1.388161e-08 1.264251e-09
#> [4,] 1.305257e-08 4.352967e-11 2.099124e-09 7.138366e-11 1.186050e-08
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 2.050880e-08 4.527687e-07 1.749657e-07 5.511505e-07 6.401039e-08
#> [2,] 7.716513e-09 1.650784e-09 7.632079e-07 9.573791e-11 3.283831e-09
#> [3,] 2.491210e-09 1.513388e-07 9.638559e-09 2.867783e-09 3.012155e-07
#> [4,] 1.088327e-08 5.566767e-08 4.772521e-08 1.975953e-09 7.811964e-10
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 9.217030e-08 3.789827e-10 3.087932e-07 8.040282e-11 1.194617e-10
#> [2,] 4.303095e-09 1.439354e-06 1.113477e-07 2.123227e-07 3.792170e-11
#> [3,] 1.152701e-07 4.269154e-10 1.741454e-09 3.029647e-09 2.290334e-08
#> [4,] 6.158134e-10 1.366739e-10 2.515163e-07 2.454828e-08 9.619810e-07
#> 
#> $Model
#> [1] "EFSAMV"
#> 
#> $Population
#> [1] 10
#>