Microbial reduction due to cooking at home

The fvCooking() function simulates the reduction in L. monocytogenes numbers in defrosted (or still frozen) vegetables due to cooking. The function can represent a scenario where the consumer gives the non-RTE frozen vegetables a non-intended use such as direct (uncooked) consumption in salads, smoothies, etc. In that case, the probability of cooking in the argument should be set to a value different from one. The variability in the effect of cooking defrosted (or frozen) vegetables is represented by a triangular distribution, as assumed in EFSA BIOHAZ (Koutsoumanis et al. (2020) .

Usage

fvCooking(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  pCooked = 0.825,
  minCook = 1,
  modeCook = 5,
  maxCook = 9
)

Arguments

data

a list of:

N

(CFU) A matrix containing the numbers of L. monocytogenes per portion of defrosted or frozen vegetables.

ProbUnitPos

Lot-specific probability of contaminated portions or servings, defrosted or not (vector).

P

Mean prevalence of contaminated portions of defrosted or frozen vegetables.

nLots

Number of lots sampled or size of the Monte Carlo simulation (scalar).

sizeLot

Number of units or portions produced in a lot (scalar).

pCooked

Probability of cooking the defrosted or frozen vegetables for consumption

minCook

(log10) Minimum value of the triangular distribution representing the variability in the reduction of L. monocytogenes in the event of cooking

modeCook

(log10) Mode value of the triangular distribution representing the variability in the reduction of L. monocytogenes in the event of cooking

maxCook

(log10) Maximum value of the triangular distribution representing the variability in the reduction of L. monocytogenes in the event of cooking

Value

A list of three elements:

N

(CFU) A matrix of size nLots lots by sizeLot units containing the numbers of L. monocytogenes in the portions or servings to be consumed;

ProbUnitPos

Lot-specific probability of contaminated portions to be consumed (vector);

P

Mean prevalence of contaminated portions or servings to be consumed (scalar).

Note

To represent variability in the effect of cooking defrosted (or frozen) vegetables, EFSA BIOHAZ (Koutsoumanis et al. (2020) assumed a triangular distribution, with parameters \(minCook=1\), \(modeCook=5\) and \(maxCook=9\). According to FSAI (2022) , 84-89% of the respondants recognised that frozen vegetables need to be cooked prior to consumption, whereasWillis et al. (2020) found that 19% of the non-RTE frozen vegetables packages have no indication for cooking. An average value for the probability of cooking of \(p\_cooked=0.825\) can be used in this function.

References

Pouillot R, Delignette-Muller M (2010). “Evaluating variability and uncertainty in microbial quantitative risk assessment using two R packages.” International Journal of Food Microbiology, 142(3), 330-40.

EFSA BIOHAZ (Koutsoumanis K, Alvarez-Ordonez A, Bolton D, Bover-Cid S, Chemaly M, Davies R, De Cesare A, Herman L, Hilbert F, Lindqvist R, Nauta M, Peixe L, Ru G, Simmons M, Skandamis P, Suffredini E, Jordan K, Sampers I, Wagner M, Felicio MTDS, Georgiadis M, Messens W, Mosbach-Schulz O, Allende A) (2020). “The public health risk posed by Listeria monocytogenes in frozen fruit and vegetables including herbs, blanched during processing.” EFSA Journal, 18(4), 1-102. doi:10.2903/j.efsa.2020.6092 , cited By 17.

Willis C, McLauchlin J, Aird H, Amar C, Barker C, Dallman T, Elviss N, Lai S, Sadler-Reeves L (2020). “Occurrence of Listeria and Escherichia coli in frozen fruit and vegetables collected from retail and catering premises in England 2018–2019.” International Journal of Food Microbiology, 334, 108849. https://www.sciencedirect.com/science/article/pii/S0168160520303433.

FSAI (2022). “National Microbiological Survey and Consumer Habits in relation to Frozen Vegetables, Fruits and Herbs (19NS6).” Food Safety Authority of Ireland. https://www.fsai.ie/Micro_Safefood_Veg_Fruit_Herbs/.

Author

Laurent Guillier

Examples

library(mc2d)
# N0 <- list(N=matrix(stats::rpois(2000, 15),
#            ncol=200, nrow=10),
#            P=0.12,
#            ProbUnitPos = 0.7)
#
# N_postcooking=fvCooking(N0,
#                          pCooked=0.825,
#                          minCook=1,
#                          modeCook=5,
#                          maxCook=9)
# hist(N_postcooking$N)