qraLm: Functions to develop quantitative risk assessment for Listeria monocytogenes in foods

qraLm: A R package for quantitative risk assessment for Listeria monocytogenes in foods.

See also

Author

Maintainer: Vasco Cadavez vcadavez@ipb.pt (ORCID)

Authors:

Régis Pouillot rpouillot.work@gmail.com (ORCID)

Laurent Guillier Laurent.guillier@anses.fr (ORCID)

Moez Sanaa sanaam@who.int (ORCID)

Ursula Gonzales-Barron ubarron@ipb.pt (ORCID)