Sampling of the RTE smoked fish characteristics

The function sfCharacteristics() samples from multiple Pert distributions the intrinsic and extrinsic characteristics of RTE smoked seafood, which affects microbial growth; namely, water activity, pH, contents of salt, phenol, nitrites, acetic acid, benzoic acid, citric acid, diacetate, lactic acid and sorbic acid, and concentration of CO_2 in equilibrium in the package. Since sampling is carried out at lot level; all the RTE seafood units of a lot have the same characteristics.

Usage

sfCharacteristics(
  nLots = data$nLots,
  awminSF = NULL,
  awmodeSF = NULL,
  awmaxSF = NULL,
  NaClminSF = 1.5,
  NaClmodeSF = 3.4,
  NaClmaxSF = 5.3,
  PminSF = 5,
  PmodeSF = 10,
  PmaxSF = 22,
  pHminSF = 5.8,
  pHmodeSF = 6.1,
  pHmaxSF = 6.5,
  CO2equilibriumminSF = 0.25,
  CO2equilibriummodeSF = 0.25,
  CO2equilibriummaxSF = 0.3,
  NITminSF = 0,
  NITmodeSF = 0,
  NITmaxSF = 0,
  aaWphminSF = 0,
  aaWphmodeSF = 0,
  aaWphmaxSF = 0,
  baWphminSF = 0,
  baWphmodeSF = 0,
  baWphmaxSF = 0,
  caWphminSF = 0,
  caWphmodeSF = 0,
  caWphmaxSF = 0,
  daWphminSF = 500,
  daWphmodeSF = 1500,
  daWphmaxSF = 1900,
  laWphminSF = 6000,
  laWphmodeSF = 12000,
  laWphmaxSF = 28000,
  saWphminSF = 0,
  saWphmodeSF = 0,
  saWphmaxSF = 0
)

Arguments

nLots

Number of lots or size of the Monte Carlo simulation

awminSF

Minimum water activity of RTE. If null,it will be calculated from NaCl concentration.

awmodeSF

Mode of water activity of RTE.

awmaxSF

Maximum water activity of RTE.

NaClminSF

(%) Minimum NaCl of RTE

NaClmodeSF

(%) Mode of NaCl of RTE

NaClmaxSF

(%) Maximum NaCl of RTE

PminSF

(ppm) Minimum phenol concentration in RTE

PmodeSF

(ppm) Mode of phenol concentration in RTE

PmaxSF

(ppm) Maximum phenol concentration in RTE

pHminSF

Minimum pH of RTE

pHmodeSF

Mode of pH of RTE

pHmaxSF

Maximum pH of RTE

CO2equilibriumminSF

(proportion) Minimum CO_2 concentration in atmosphere in RTE package (e.g. .0.25)

CO2equilibriummodeSF

(proportion) Mode of CO_2 concentration in atmosphere in RTE package

CO2equilibriummaxSF

(proportion) Maximum CO_2 concentration in atmosphere in RTE package

NITminSF

(ppm) Minimum nitrites concentration in RTE

NITmodeSF

(ppm) Mode of nitrites concentration in RTE

NITmaxSF

(ppm) Maximum nitrites concentration in RTE

aaWphminSF

(ppm) Minimum acetic acid concentration in RTE

aaWphmodeSF

(ppm) Mode of acetic acid concentration in RTE

aaWphmaxSF

(ppm) Maximum acetic acid concentration in RTE

baWphminSF

(ppm) Minimum benzoic acid concentration in RTE

baWphmodeSF

(ppm) Mode benzoic acid concentration in RTE

baWphmaxSF

(ppm) Maximum benzoic acid concentration in RTE

caWphminSF

(ppm) Minimum citric acid concentration in RTE

caWphmodeSF

(ppm) Mode of citric acid concentration in RTE

caWphmaxSF

(ppm) Maximum citric acid concentration in RTE

daWphminSF

(ppm) Minimum diacetate concentration in RTE

daWphmodeSF

(ppm) Mode of diacetate concentration in RTE

daWphmaxSF

(ppm) Maximum diacetate concentration in RTE

laWphminSF

(ppm) Minimum lactic acid concentration in RTE

laWphmodeSF

(ppm) Mode of lactic acid concentration in RTE

laWphmaxSF

(ppm) Maximum lactic acid concentration in RTE

saWphminSF

(ppm) Minimum sorbic acid concentration in RTE

saWphmodeSF

(ppm) Mode of sorbic acid concentration in RTE

saWphmaxSF

(ppm) Maximum sorbic acid concentration in RTE

Value

A list containing vectors aw, NaCl, pH, P, CO2equi, NIT, aaWph, baWph, caWph, daWph, laWph, saWph to be used in the growth functions sfMejlholmDalgaard() and sfMejlholmDalgaardLAB() for estimating specific growth rates of L. monocytogenes and lactic acid bacteria, respectively, in RTE seafood.

Note

All parameters end with SF not to be confounded with growth parameters. If parameters for water activity aw are NULL, aw will be evaluated from NaCl concentration in the growth functions. The default parameters of the Pert distributions describing variability in pH, contents of salt, phenol, diacetate, lactic acid, and concentration of CO_2 in equilibrium for smoked seafood, were taken from WHO (2022) (Table 3.25) and Pérez-Rodríguez (2017) .

References

WHO (2022). “A Roadmap for the Development of Risk Assessment Models of Listeria monocytogenes in Selected Produce and Seafood Products.” World Health Organization. Pérez-Rodríguez FCEBSJAVA (2017). “Closing gaps for performing a risk assessment on Listeria monocytogenes in ready-to-eat (RTE) foods: activity 2, a quantitative risk characterization on L. monocytogenes in RTE foods; starting from the retail stage.” EFSA Supporting publication 2017:EN-1252. Pouillot R, Delignette-Muller M (2010). “Evaluating variability and uncertainty in microbial quantitative risk assessment using two R packages.” International Journal of Food Microbiology, 142(3), 330-40.

Author

Regis Pouillot rpouillot.work@gmail.com

Examples

N <- matrix(round(10^rnorm(100 * 100, 0, 3)),
  ncol = 100, nrow = 100
)

dat <- list(N = N, nLots = 1000)
dat <- sfCharacteristics(dat)
str(dat)
#> List of 12
#>  $ aw     : NULL
#>  $ NaCl   : num [1:2] 2.18 2.18
#>  $ pH     : num [1:2] 6.09 5.95
#>  $ P      : num [1:2] 15.5 12.7
#>  $ CO2equi: num [1:2] 0.275 0.264
#>  $ NIT    : num [1:2] 0 0
#>  $ aaWph  : num [1:2] 0 0
#>  $ baWph  : num [1:2] 0 0
#>  $ caWph  : num [1:2] 0 0
#>  $ daWph  : num [1:2] 1111 1557
#>  $ laWph  : num [1:2] 13473 10225
#>  $ saWph  : num [1:2] 0 0