Dose-response model function for the qraLm models

This function provides the marginal probability of invasive listeriosis in a given population for a given Dose in CFU using the JEMRA, the Pouillot, the Fritsch or the EFSA dose-response models or the model developed within this project (EFSAMV,EFSAV,EFSALV) (see References).

Usage

DRForModel(data = list(), model = "JEMRA", population = 1, Poisson = FALSE)

Arguments

data

a list of (a minima):

N

(CFU) A matrix containing the numbers of L. monocytogenes per portion of defrosted or frozen vegetables.

model

either JEMRA, Pouillot, Fritsch, EFSA, EFSAMV,EFSAV or EFSALV

population

considered population (scalar).

Poisson

if TRUE, assume that Dose is the mean of a Poisson distribution. (actual LogNormal Poisson). If FALSE (default), assume that Dose is the actual number of bacteria.

Value

the data object with added:

Risk

A matrix of risk, of size similar to N

Model

the Model

Population

the Population

Details

see doseresponsemodels::DR() or doseresponsemodels::DRQuick() for details on population and models.

Note

This function uses (for all model but JEMRA) a linear approximation (approxfun) from the exact doseresponsemodels::DR() model evaluated on \(Dose = c(0,10^{seq(-5,12,length=1701)})\) (if Poisson=TRUE) or \(c(0,10^{seq(0,12,length=2000)})\) (if Poisson=FALSE). Any Dose lower or higher than these ranges will lead to NA.

References

EFSA (2018). “Scientific opinion on the Listeria monocytogenes contamination of ready-to-eat foods and the risk from human health in the EU.” EFSA Journal, 16(1), 5134. FAO-WHO (2004). “Risk assessment of Listeria monocytogenes in ready-to-eat foods: Technical report.” World Health Organization and Food and Agriculture Organization of the United Nations. Fritsch L, Guillier L, Augustin J (2018). “Next generation quantitative microbiological risk assessment: Refinement of the cold smoked salmon-related listeriosis risk model by integrating genomic data.” Microbial Risk Analysis, 10, 20--27. doi:10.1016/j.mran.2018.06.003 . Pouillot R, Hoelzer K, Chen Y, Dennis SB (2015). “Listeria monocytogenes dose response revisited--incorporating adjustments for variability in strain virulence and host susceptibility.” Risk Analysis, 35(1), 90--108. doi:10.1111/risa.12235 .

Author

Regis Pouillot, Vasco Cadavez

Examples

data <- list(N = matrix(10^runif(100, 0, 5), ncol = 25))
DRForModel(data, "JEMRA", 1)
#> $N
#>              [,1]       [,2]         [,3]       [,4]         [,5]       [,6]
#> [1,]     2.533674   1.088923  4617.532225   1.483224   104.157510   28.14890
#> [2,] 14847.905209 214.766968  7287.986440  39.987905     2.081169 2465.48371
#> [3,]  1008.798508 308.238543 23604.961899 102.716732    87.802436 4748.95434
#> [4,]     6.110014  28.108406     7.493818   9.513694 75463.887875    9.54517
#>              [,7]        [,8]      [,9]        [,10]     [,11]        [,12]
#> [1,] 79927.894586 3013.834464   31.9267  3405.408289 2516.6023     3.020792
#> [2,]  5100.440032 2771.915691 1521.6966 55320.090453  312.0528  6729.779266
#> [3,]     1.808137    1.432688  248.3835     7.974323 1615.8340  7053.002235
#> [4,]   447.777552   13.421859  144.8293    12.147849 2001.8049 89858.995903
#>            [,13]       [,14]       [,15]       [,16]        [,17]     [,18]
#> [1,] 71220.41737    7.471013   10.492808 13365.86607   957.886261 528.46590
#> [2,]    88.29046  454.849481 3689.660396    23.39329     9.069642  24.71643
#> [3,]   202.27040  293.890184    2.118754   708.31225 54804.937750 171.20338
#> [4,]    37.68849 7880.306869   59.024728    47.70846   515.705833  72.03705
#>           [,19]       [,20]      [,21]      [,22]        [,23]       [,24]
#> [1,]   1.381354    76.71047  436.09909   251.4061  2284.725761   391.54635
#> [2,] 213.764683   630.32452 1007.36676 39813.2592     1.963679 17005.92553
#> [3,]  89.157306 19293.80702   20.27014   100.8326 96682.011294  3902.55130
#> [4,]   1.259871    83.95538   28.20011    11.6376     5.561313    16.09052
#>             [,25]
#> [1,]   543.521643
#> [2,] 14928.256070
#> [3,]     1.379686
#> [4,]   222.290805
#> 
#> $Risk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 5.995204e-14 2.575717e-14 1.091942e-10 3.508305e-14 2.463141e-12
#> [2,] 3.511196e-10 5.078715e-12 1.723445e-10 9.455770e-13 4.918288e-14
#> [3,] 2.385581e-11 7.289169e-12 5.582043e-10 2.429057e-12 2.076339e-12
#> [4,] 1.444400e-13 6.646905e-13 1.771916e-13 2.249312e-13 1.784551e-09
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 6.656897e-13 1.890115e-09 7.127043e-11 7.549517e-13 8.053025e-11
#> [2,] 5.830314e-11 1.206140e-10 6.554957e-11 3.598466e-11 1.308196e-09
#> [3,] 1.123021e-10 4.274359e-14 3.386180e-14 5.873746e-12 1.886269e-13
#> [4,] 2.257083e-13 1.058897e-11 3.174128e-13 3.424927e-12 2.872147e-13
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 5.951195e-11 7.138734e-14 1.684203e-09 1.766365e-13 2.481348e-13
#> [2,] 7.379319e-12 1.591441e-10 2.087885e-12 1.075617e-11 8.725221e-11
#> [3,] 3.821088e-11 1.667876e-10 4.783285e-12 6.949885e-12 5.007106e-14
#> [4,] 4.733813e-11 2.124963e-09 8.912870e-13 1.863515e-10 1.395772e-12
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 3.160726e-10 2.265188e-11 1.249700e-11 3.264056e-14 1.813993e-12
#> [2,] 5.532241e-13 2.144951e-13 5.845324e-13 5.055067e-12 1.490574e-11
#> [3,] 1.675005e-11 1.296014e-09 4.048539e-12 2.108425e-12 4.562551e-10
#> [4,] 1.128209e-12 1.219524e-11 1.703526e-12 2.975398e-14 1.985301e-12
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 1.031275e-11 5.945244e-12 5.402867e-11 9.259149e-12 1.285305e-11
#> [2,] 2.382194e-11 9.414940e-10 4.640732e-14 4.021519e-10 3.530197e-10
#> [3,] 4.793943e-13 2.384426e-12 2.286312e-09 9.228651e-11 3.264056e-14
#> [4,] 6.669110e-13 2.752243e-13 1.315614e-13 3.804734e-13 5.256684e-12
#> 
#> $lotMeanRisk
#> [1] 1.737917e-10 1.650660e-10 2.034299e-10 1.674120e-10
#> 
#> $servingRisk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 5.995204e-14 2.575717e-14 1.091942e-10 3.508305e-14 2.463141e-12
#> [2,] 3.511196e-10 5.078715e-12 1.723445e-10 9.455770e-13 4.918288e-14
#> [3,] 2.385581e-11 7.289169e-12 5.582043e-10 2.429057e-12 2.076339e-12
#> [4,] 1.444400e-13 6.646905e-13 1.771916e-13 2.249312e-13 1.784551e-09
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 6.656897e-13 1.890115e-09 7.127043e-11 7.549517e-13 8.053025e-11
#> [2,] 5.830314e-11 1.206140e-10 6.554957e-11 3.598466e-11 1.308196e-09
#> [3,] 1.123021e-10 4.274359e-14 3.386180e-14 5.873746e-12 1.886269e-13
#> [4,] 2.257083e-13 1.058897e-11 3.174128e-13 3.424927e-12 2.872147e-13
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 5.951195e-11 7.138734e-14 1.684203e-09 1.766365e-13 2.481348e-13
#> [2,] 7.379319e-12 1.591441e-10 2.087885e-12 1.075617e-11 8.725221e-11
#> [3,] 3.821088e-11 1.667876e-10 4.783285e-12 6.949885e-12 5.007106e-14
#> [4,] 4.733813e-11 2.124963e-09 8.912870e-13 1.863515e-10 1.395772e-12
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 3.160726e-10 2.265188e-11 1.249700e-11 3.264056e-14 1.813993e-12
#> [2,] 5.532241e-13 2.144951e-13 5.845324e-13 5.055067e-12 1.490574e-11
#> [3,] 1.675005e-11 1.296014e-09 4.048539e-12 2.108425e-12 4.562551e-10
#> [4,] 1.128209e-12 1.219524e-11 1.703526e-12 2.975398e-14 1.985301e-12
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 1.031275e-11 5.945244e-12 5.402867e-11 9.259149e-12 1.285305e-11
#> [2,] 2.382194e-11 9.414940e-10 4.640732e-14 4.021519e-10 3.530197e-10
#> [3,] 4.793943e-13 2.384426e-12 2.286312e-09 9.228651e-11 3.264056e-14
#> [4,] 6.669110e-13 2.752243e-13 1.315614e-13 3.804734e-13 5.256684e-12
#> 
#> $Model
#> [1] "JEMRA"
#> 
#> $Population
#> [1] 1
#> 
DRForModel(data, "EFSAMV", 10)
#> $N
#>              [,1]       [,2]         [,3]       [,4]         [,5]       [,6]
#> [1,]     2.533674   1.088923  4617.532225   1.483224   104.157510   28.14890
#> [2,] 14847.905209 214.766968  7287.986440  39.987905     2.081169 2465.48371
#> [3,]  1008.798508 308.238543 23604.961899 102.716732    87.802436 4748.95434
#> [4,]     6.110014  28.108406     7.493818   9.513694 75463.887875    9.54517
#>              [,7]        [,8]      [,9]        [,10]     [,11]        [,12]
#> [1,] 79927.894586 3013.834464   31.9267  3405.408289 2516.6023     3.020792
#> [2,]  5100.440032 2771.915691 1521.6966 55320.090453  312.0528  6729.779266
#> [3,]     1.808137    1.432688  248.3835     7.974323 1615.8340  7053.002235
#> [4,]   447.777552   13.421859  144.8293    12.147849 2001.8049 89858.995903
#>            [,13]       [,14]       [,15]       [,16]        [,17]     [,18]
#> [1,] 71220.41737    7.471013   10.492808 13365.86607   957.886261 528.46590
#> [2,]    88.29046  454.849481 3689.660396    23.39329     9.069642  24.71643
#> [3,]   202.27040  293.890184    2.118754   708.31225 54804.937750 171.20338
#> [4,]    37.68849 7880.306869   59.024728    47.70846   515.705833  72.03705
#>           [,19]       [,20]      [,21]      [,22]        [,23]       [,24]
#> [1,]   1.381354    76.71047  436.09909   251.4061  2284.725761   391.54635
#> [2,] 213.764683   630.32452 1007.36676 39813.2592     1.963679 17005.92553
#> [3,]  89.157306 19293.80702   20.27014   100.8326 96682.011294  3902.55130
#> [4,]   1.259871    83.95538   28.20011    11.6376     5.561313    16.09052
#>             [,25]
#> [1,]   543.521643
#> [2,] 14928.256070
#> [3,]     1.379686
#> [4,]   222.290805
#> 
#> $Risk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 6.124864e-11 2.632346e-11 1.116234e-07 3.585523e-11 2.517888e-09
#> [2,] 3.589297e-07 5.191744e-09 1.761784e-07 9.666616e-10 5.030986e-11
#> [3,] 2.438654e-08 7.451311e-09 5.706194e-07 2.483059e-09 2.122523e-09
#> [4,] 1.477026e-10 6.794884e-10 1.811544e-10 2.299826e-10 1.824218e-06
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 6.804673e-10 1.932126e-06 7.285592e-08 7.717912e-10 8.232175e-08
#> [2,] 5.960019e-08 1.232971e-07 6.700782e-08 3.678525e-08 1.337281e-06
#> [3,] 1.148004e-07 4.370963e-11 3.463358e-11 6.004385e-09 1.927701e-10
#> [4,] 2.307435e-10 1.082451e-08 3.244580e-10 3.501080e-09 2.936603e-10
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 6.083592e-08 7.302416e-11 1.721641e-06 1.806031e-10 2.536515e-10
#> [2,] 7.543516e-09 1.626844e-07 2.134320e-09 1.099546e-08 8.919321e-08
#> [3,] 3.906091e-08 1.704979e-07 4.889654e-09 7.104456e-09 5.121844e-11
#> [4,] 4.839130e-08 2.172189e-06 9.110759e-10 1.904970e-07 1.426855e-09
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 3.231033e-07 2.315579e-08 1.277505e-08 3.339263e-11 1.854387e-09
#> [2,] 5.655057e-10 2.192482e-10 5.974913e-10 5.167515e-09 1.523737e-08
#> [3,] 1.712263e-08 1.324828e-06 4.138644e-09 2.155275e-09 4.664033e-07
#> [4,] 1.153297e-09 1.246659e-08 1.741413e-09 3.045593e-11 2.029525e-09
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 1.054219e-08 6.077453e-09 5.523058e-08 9.465181e-09 1.313901e-08
#> [2,] 2.435193e-08 9.624301e-07 4.746969e-11 4.110969e-07 3.608721e-07
#> [3,] 4.900073e-10 2.437512e-09 2.337120e-06 9.433959e-08 3.335231e-11
#> [4,] 6.817052e-10 2.813257e-10 1.344383e-10 3.889698e-10 5.373624e-09
#> 
#> $lotMeanRisk
#> [1] 1.776553e-07 1.687370e-07 2.079524e-07 1.711331e-07
#> 
#> $servingRisk
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 6.124864e-11 2.632346e-11 1.116234e-07 3.585523e-11 2.517888e-09
#> [2,] 3.589297e-07 5.191744e-09 1.761784e-07 9.666616e-10 5.030986e-11
#> [3,] 2.438654e-08 7.451311e-09 5.706194e-07 2.483059e-09 2.122523e-09
#> [4,] 1.477026e-10 6.794884e-10 1.811544e-10 2.299826e-10 1.824218e-06
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 6.804673e-10 1.932126e-06 7.285592e-08 7.717912e-10 8.232175e-08
#> [2,] 5.960019e-08 1.232971e-07 6.700782e-08 3.678525e-08 1.337281e-06
#> [3,] 1.148004e-07 4.370963e-11 3.463358e-11 6.004385e-09 1.927701e-10
#> [4,] 2.307435e-10 1.082451e-08 3.244580e-10 3.501080e-09 2.936603e-10
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 6.083592e-08 7.302416e-11 1.721641e-06 1.806031e-10 2.536515e-10
#> [2,] 7.543516e-09 1.626844e-07 2.134320e-09 1.099546e-08 8.919321e-08
#> [3,] 3.906091e-08 1.704979e-07 4.889654e-09 7.104456e-09 5.121844e-11
#> [4,] 4.839130e-08 2.172189e-06 9.110759e-10 1.904970e-07 1.426855e-09
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 3.231033e-07 2.315579e-08 1.277505e-08 3.339263e-11 1.854387e-09
#> [2,] 5.655057e-10 2.192482e-10 5.974913e-10 5.167515e-09 1.523737e-08
#> [3,] 1.712263e-08 1.324828e-06 4.138644e-09 2.155275e-09 4.664033e-07
#> [4,] 1.153297e-09 1.246659e-08 1.741413e-09 3.045593e-11 2.029525e-09
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 1.054219e-08 6.077453e-09 5.523058e-08 9.465181e-09 1.313901e-08
#> [2,] 2.435193e-08 9.624301e-07 4.746969e-11 4.110969e-07 3.608721e-07
#> [3,] 4.900073e-10 2.437512e-09 2.337120e-06 9.433959e-08 3.335231e-11
#> [4,] 6.817052e-10 2.813257e-10 1.344383e-10 3.889698e-10 5.373624e-09
#> 
#> $Model
#> [1] "EFSAMV"
#> 
#> $Population
#> [1] 10
#>