Generation of contaminated lots with variability in prevalence and concentration

The function Lot2LotGen() generates a matrix of contaminated lots from information on the parameters of a beta distribution (alpha, beta) representing the between-lot variability in prevalence; the parameters of a normal distribution (mean, sd) representing the total variability in the microbial concentration in the contaminated units (log10 CFU/g); and the proportion of such total variance assigned to between-lot variance. Every row of the matrix corresponds to a production lot, broken down in units or portions, which are the columns of the matrix.

Usage

Lot2LotGen(
  nLots,
  sizeLot,
  unitSize,
  betaAlpha,
  betaBeta,
  C0MeanLog,
  C0SdLog,
  propVarInter,
  Poisson = FALSE,
  ...
)

Arguments

nLots

Number of lots sampled or size of the Monte Carlo simulation (scalar).

sizeLot

Number of units or portions produced in a lot (scalar).

unitSize

Weight (g) of single unit or portion from a lot (scalar).

betaAlpha

Parameter alpha of the beta distribution representing the between-batch variability in prevalence (scalar).

betaBeta

Parameter beta of the beta distribution representing the between-batch variability in prevalence (scalar).

C0MeanLog

Mean parameter (log10 CFU/g) of the normal distribution representing the total variability in the microbial concentration of contaminated units (scalar or vector).

C0SdLog

Standard deviation parameter (log10 CFU/g) of the normal distribution representing the total variability in the microbial concentration of contaminated units (scalar or vector).

propVarInter

proportion of the total variance (\(C0\_sdLog10^2\)) attributed to between-lot variance. The remaining (1-propVarInter) is the proportion of within-lot variance.

Poisson

need to be defined

...

Other options used to control LotGen()

Value

A list of four elements:

N

(CFU) A matrix of size nLots lots by sizeLot units representing the microbial numbers in the units or portions from contaminated lots;

ProbUnitPos

Probability that the lot is contaminated (a lot is considered contaminated if at least one unit or portion is contaminated) (vector);

P

Mean prevalence of contaminated lots (scalar);

betaGen

True prevalence of contaminated units in the lots (sampled from the beta distribution) (vector).

Note

This function requires that the variability in the prevalence of L. monocytogenes between lots is represented by a beta distribution. To determine the parameters of such a beta distribution, a beta-binomial distribution was fitted to survey data (compiled in Table 3.21 of Background2022;textualqraLm, consisting of positive samples and total sample size. The beta-binomial was fitted through a Markov-Chain Monte Carlo approach, with gamma(0.01, 0.01) as priors for both parameters, and assuming that every data source (a published paper) represents a different lot. The parameters of the beta distribution were determined at: betaAlpha=0.5112 and betaBeta=9.959. The mean (C0MeanLog=1.023 log10 CFU/g) and standard deviation (C0SdLog=0.3267 log10\ CFU/g) of the normal distribution representing the overall variability in L. monocytogenes were obtained using data from Jeya2011;textualqraLm (3.0 MPN/g, 3.6 MPN/g in freshly harvested vegetables), Magdovitz2021;textualqraLm (21, 36 and 75 MPN/g in corn and peas arriving at the frozen food facility), and Kuan2017;textualqraLm (3.0 MPN/g in bulk carrots at retail). The algorithm decomposes at the log10 level the total variance in concentration into between-lot variance and within-lot variance according to the proportion propVarInter.

References

statsqraLm Background2022qraLm Jeya2011qraLm Magdovitz2021qraLm Kuan2017qraLm

See also

Author

Vasco Cadavez vcadavez@ipb.pt and Regis Pouillot rpouillot.work@gmail.com

Examples

dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)

print(dat$P)
#> [1] 0.7103589
summary(c(dat$N))
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     0.0     0.0     0.0   417.9     0.0 59206.0