Generation of contaminated lots

The function LotGen() generates a matrix of contaminated lots from information on the parameters of a normal distribution (mean, sd) representing the microbial concentration in the contaminated units (log10 CFU/g). Every row of the matrix corresponds to a production lot, broken down in units or portions, which are the columns of the matrix. There is no assumption that lots can have different prevalence values, and therefore the microbial numbers of all units or portions produced in a lot are sampled from a microbial concentration distribution.

Usage

LotGen(nLots, sizeLot, P, C0MeanLog, C0SdLog, unitSize, ...)

Arguments

nLots

number of lots sampled or size of the Monte Carlo simulation (scalar).

sizeLot

number of units or portions produced in a lot (scalar).

P

prevalence of contaminated lots (scalar)

C0MeanLog

(log10 CFU/g) mean parameter of the normal distribution representing the variability in the microbial concentration of contaminated units (scalar or vector).

C0SdLog

(log10 CFU/g) standard deviation parameter of the normal distribution representing the variability in the microbial concentration of contaminated units (scalar or vector).

unitSize

(g) weight of single unit or portion from a lot (scalar).

...

Other options used to control LotGen()

Value

A list of two elements:

N

(CFU) A matrix of size nLots lots by sizeLot units representing the microbial numbers in the units or portions from contaminated lots;

P

Prevalence of contaminated lots (scalar).

Note

For the prevalence of lots contaminated with L. monocytogenes, \(P=0.057\) is suggested from the information published in FSAI2022;textualqraLm, which detected L. monocytogenes in 21 out of 366 in non-RTE frozen vegetables. The mean (\(C0\_meanLog10=1.023\ log10\ CFU/g\)) and standard deviation (\(C0\_sdLog10=0.3267\ log10\ CFU/g\)) of the normal distribution representing the variability in L. monocytogenes were obtained using data from Jeya2011;textualqraLm (\(3.0\ MPN/g\), \(3.6\ MPN/g\) in freshly harvested vegetables), Magdovitz2021;textualqraLm (21, 36 and 75 MPN/g in corn and peas arriving at the frozen food facility), and Kuan2017;textualqraLm (\(3.0\ MPN/g\) in bulk carrots at retail).

References

statsqraLm

FSAI2022qraLm

Jeya2011qraLm

Magdovitz2021qraLm

Kuan2017qraLm

See also

Author

Vasco Cadavez vcadavez@ipb.pt

Examples

lots <- LotGen(
               nLots = 500,
               sizeLot = 200,
               P = 0.057,
               C0MeanLog = 1.023,
               C0SdLog = 0.3267,
               unitSize = 500
               )
str(lots)
#> List of 3
#>  $ N               : num [1:500, 1:200] 1004 12218 3626 5426 4694 ...
#>  $ P               : num 0.057
#>  $ LotGenParameters:List of 6
#>   ..$ nLots    : num 500
#>   ..$ sizeLot  : num 200
#>   ..$ P        : num 0.057
#>   ..$ C0MeanLog: num 1.02
#>   ..$ C0SdLog  : num 0.327
#>   ..$ unitSize : num 500
#>  - attr(*, "class")= chr [1:2] "list" "qraLm"
summary(c(lots$N))
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     183    3171    5290    7007    8805  156868