Route of contamination of cantaloupe from irrigation water to rind

The function caIrrig2rind() evaluates the contamination of cantaloupes through irrigation water only. It considers water contamination characteristics: prevalence pIrrig and concentration cIrrig. pIrrig and cIrrig have to be chosen by the user according to existing data of prevalence. cIrrig is conditional to water sources contaminated with L. monocytogenes.

Usage

caIrrig2rind(
  nLots,
  sizeLot,
  pIrrig = 0.131,
  cIrrigLogMin = -1.52,
  cIrrigLogMax = 1.04,
  cantaWeight = 1000,
  pWaterGainMin = 0,
  pWaterGainMax = 0.004
)

Arguments

nLots

Size of the Monte Carlo simulation (scalar).

sizeLot

Number of cantaloupes per cultivation lot or field (scalar).

pIrrig

Prevalence of contamination in irrigation water (provided by the user, \(default=0.131\) according to Raschle2021;textualqraLm).

cIrrigLogMin

(log CFU/L) Minimum value of the uniform distribution (provided by the user, \(default=-1.52\) according to Sharma2020;textualqraLm).

cIrrigLogMax

(log CFU/L) Maximum value of the uniform distribution (provided by the user, \(default=1.04\) according to Sharma2020;textualqraLm).

cantaWeight

(g) Weight of a cantaloupe.

pWaterGainMin

Minimum value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\(default=0.0\)).

pWaterGainMax

Maximum value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\(default=0.004\) according to Richards2004;textualqraLm).

Value

A list of two elements:

N

(CFU) A matrix of size nLots by sizeLot containing the numbers of L. monocytogenes cells on cantaloupe from the irrigation water route;

P

Prevalence of field lots of cantaloupes contaminated with L. monocytogenes from the irrigation water route.

Note

Background2022;textualqraLm lists many estimates of L. monocytogenes in water environments. The estimate of \(0.131\) provided by Raschle2021;textualqraLm has been chosen as default. The distribution about the concentration of L. monocytogenes in irrigation water is represented as a uniform distribution, using the minimum and maximum values from Sharma2020;textualqraLm, who reported \(<0.03\ to\ 11\) MPN L. monocytogenes/ L water. The algorithm assumes that the amount of water deposited on the cantaloupe rind after the last irrigation is as a percentage of the cantaloupe weight, and is sampled from a uniform distribution of parameters between a minimum value of zero and a maximum value of \(0.004\), taken from Richards2004;textualqraLm.

References

statsqraLm

mc2dqraLm

Background2022qraLm

Raschle2021qraLm

Richards2004qraLm

Sharma2020qraLm

Author

Laurent Guillier

Examples

dat <- caPrimaryProduction(
  nLots = 100,
  sizeLot = 100,
  pSoil = 0.089,
  pManure = 0.1,
  pIrrigRaining = 0.05,
  pFoil = 0.5,
  rFoil = 0.75,
  pIrrig = 0.131,
  cantaWeight = 800
)

caIrrig2rind(
  nLots = 5,
  sizeLot = 10,
  pIrrig = 0.5,
  cIrrigLogMin = 1,
  cIrrigLogMax = 2
)
#> $N
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    0    1    0    0    0    0    0    0    0     0
#> [2,]    0    0    0    0    0    0    0    0    0     1
#> [3,]    0    0    0    0    0    0    0    0    0     1
#> [4,]    0    0    0    0    0    0    0    0    0     1
#> [5,]    0    0    0    0    0    0    0    0    0     1
#> 
#> $P
#> [1] 0.2
#>