Route of contamination of cantaloupe from irrigation water to rind

The function caIrrig2rind() evaluates the contamination of cantaloupes through irrigation water only. It considers water contamination characteristics: prevalence pIrrig and concentration cIrrig. pIrrig and cIrrig have to be chosen by the user according to existing data of prevalence. cIrrig is conditional to water sources contaminated with L. monocytogenes.

Usage

caIrrig2rind(
  nLots,
  sizeLot,
  pIrrig = 0.131,
  cIrrigLogMin = -1.52,
  cIrrigLogMax = 1.04,
  cantaWeight = 1000,
  pWaterGainMin = 0,
  pWaterGainMax = 0.004
)

Arguments

nLots

Size of the Monte Carlo simulation (scalar).

sizeLot

Number of cantaloupes per cultivation lot or field (scalar).

pIrrig

Prevalence of contamination in irrigation water (provided by the user, \(default=0.131\) according to Raschle et al. (2021) ).

cIrrigLogMin

(log CFU/L) Minimum value of the uniform distribution (provided by the user, \(default=-1.52\) according to Sharma et al. (2020) ).

cIrrigLogMax

(log CFU/L) Maximum value of the uniform distribution (provided by the user, \(default=1.04\) according to Sharma et al. (2020) ).

cantaWeight

(g) Weight of a cantaloupe.

pWaterGainMin

Minimum value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\(default=0.0\)).

pWaterGainMax

Maximum value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\(default=0.004\) according to Richards and Beuchat (2004) ).

Value

A list of two elements:

N

(CFU) A matrix of size nLots by sizeLot containing the numbers of L. monocytogenes cells on cantaloupe from the irrigation water route;

P

Prevalence of field lots of cantaloupes contaminated with L. monocytogenes from the irrigation water route.

Note

WHO (2022) lists many estimates of L. monocytogenes in water environments. The estimate of \(0.131\) provided by Raschle et al. (2021) has been chosen as default. The distribution about the concentration of L. monocytogenes in irrigation water is represented as a uniform distribution, using the minimum and maximum values from Sharma et al. (2020) , who reported \(<0.03\ to\ 11\) MPN L. monocytogenes/ L water. The algorithm assumes that the amount of water deposited on the cantaloupe rind after the last irrigation is as a percentage of the cantaloupe weight, and is sampled from a uniform distribution of parameters between a minimum value of zero and a maximum value of \(0.004\), taken from Richards and Beuchat (2004) .

References

Team RC (2022). R: A Language and Environment for Statistical Computing. R Foundation for Statistical Computing, Vienna, Austria. https://www.R-project.org/.

Pouillot R, Delignette-Muller M (2010). “Evaluating variability and uncertainty in microbial quantitative risk assessment using two R packages.” International Journal of Food Microbiology, 142(3), 330-40.

WHO (2022). “A Roadmap for the Development of Risk Assessment Models of Listeria monocytogenes in Selected Produce and Seafood Products.” World Health Organization.

Raschle S, Stephan R, Stevens M, Cernela N, Zurfluh K, Muchaamba F, Nuesch-Inderbinen M (2021). “Environmental dissemination of pathogenic Listeria monocytogenes in flowing surface waters in Switzerland.” Scientific Reports, 11(9066).

Richards GM, Beuchat LR (2004). “Attachment of Salmonella Poona to Cantaloupe Rind and Stem Scar Tissues as Affected by Temperature of Fruit and Inoculum.” Journal of Food Protection, 67(7), 1359-1364. doi:10.4315/0362-028X-67.7.1359 .

Sharma M, Handy ET, East CL, Kim S, Jiang C, Callahan MT, Allard SM, Micallef S, Craighead S, Anderson-Coughlin B, Gartley S, Vanore A, Kniel KE, Haymaker J, Duncan R, Foust D, White C, Taabodi M, Hashem F, Parveen S, May E, Bui A, Craddock H, Kulkarni P, Murray RT, Sapkota AR (2020). “Prevalence of Salmonella and Listeria monocytogenes in non-traditional irrigation waters in the Mid-Atlantic United States is affected by water type, season, and recovery method.” PLOS ONE, 15(3), 1-15. doi:10.1371/journal.pone.0229365 .

Author

Laurent Guillier

Examples

dat <- caPrimaryProduction(
  nLots = 100,
  sizeLot = 100,
  pSoil = 0.089,
  pManure = 0.1,
  pIrrigRaining = 0.05,
  pFoil = 0.5,
  rFoil = 0.75,
  pIrrig = 0.131,
  cantaWeight = 800
)

caIrrig2rind(
  nLots = 5,
  sizeLot = 10,
  pIrrig = 0.5,
  cIrrigLogMin = 1,
  cIrrigLogMax = 2
)
#> $N
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    0    1    0    0    0    0    0    0    0     0
#> [2,]    0    0    0    0    0    0    0    0    0     1
#> [3,]    0    0    0    0    0    0    0    0    0     1
#> [4,]    0    0    0    0    0    0    0    0    0     1
#> [5,]    0    0    0    0    0    0    0    0    0     1
#> 
#> $P
#> [1] 0.2
#>