Route of contamination of cantaloupe from soil to rind

The function caSoil2rind() evaluates the contamination of cantaloupes through soil only. It considers soil contamination characteristics such as prevalence pSoil and concentration c_soil as well as the quantity of soil deposited on the cantaloupe rind. Furthermore, the algorithm allows for the use of foils to protect cantaloupe from soil contamination, and the user must provide the probability of growing cantaloupes using such protective foils. pSoil and c_soil have to be chosen by the user according to existing data; whereas pSoil is conditional to several risk factors. The identified risk factors Background2022qraLm are:

irrigation in the previous days before harvesting; and

the use of organic fertilizer.

which affect pSoil as associated odds-ratios (F_irrig_rain and fManure). In this context, the user has to define the proportion of fields practising irrigation prior to harvest or undergoing rain (p_irrig_rain), and the proportion of fields using organic amendments (pManure).

Usage

caSoil2rind(
  nLots,
  sizeLot,
  pSoil = 0.089,
  fManure = 7,
  pManure = 0.5,
  fIrrigRaining = 25,
  pIrrigRaining = 0.1,
  cSoilLogMin = -1,
  cSoilLogMode = 0.6,
  cSoilLogMax = 1.48,
  qSoilMin = 0.05,
  qSoilMode = 0.5,
  qSoilMax = 5,
  pFoil = 0.5,
  rFoil = 0.9
)

Arguments

nLots

Size of the Monte Carlo simulation (scalar).

sizeLot

Number of cantaloupes per cultivation lot or field (scalar).

pSoil

Probability of contamination of soil (provided by the user, \(default=0.089\) according to Strawn2013a;textualqraLm)

fManure

Odds-ratio estimate associated to use or organic amendment in soil (\(default=7.0\) according to Strawn2013b;textualqraLm)

pManure

Proportion of fields using organic amendments (provided by user, \(default=0.5\))

fIrrigRaining

Odds-ratio estimate associated to use of irrigation and raining events up to 2 days before harvest (\(default=25.0\) according to Weller2015;textualqraLm).

pIrrigRaining

Proportion of fields that undergo irrigation or raining just previous harvest (provided by user, \(default=0.1\)).

cSoilLogMin

(log10 CFU/g) Minimum value of the triangular distribution describing variability of concentration (according to Dowe1997;textualqraLm \(default=-1\ log10\ CFU/g\)).

cSoilLogMode

(log10 CFU/g) Mode value of the triangular distribution describing variability of concentration (according to Dowe1997;textualqraLm \(default=0.6\ log10\ CFU/g\)).

cSoilLogMax

(log10 CFU/g) Maximum value of the triangular distribution describing variability of concentration (according to Dowe1997;textualqraLm: \(default=1.48\ log10\ CFU/g\)).

qSoilMin

(g) Minimum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\(default=0.05\ g\)).

qSoilMode

(g) Mode value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\(default=0.5\ g\)).

qSoilMax

(g) Maximum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\(default=5\ g\)).

pFoil

Proportion of fields grown in foil (e.g. plastic mulch) (EKE: \(default=0.5\))

rFoil

Reduction fraction of the quantity of soil transferred to rind (EKE: \(default=0.9\))

Value

A list of two elements:

N

(CFU) A matrix of size nLots by sizeLot containing the numbers of L. monocytogenes cells on cantaloupe from the soil route;

P

Prevalence of field lots of cantaloupes contaminated with L. monocytogenes from the soil route (scalar).

Note

The prevalence in soil, pSoil, must be provided by the user according to existing data. Background2022;textualqraLm lists various prevalence estimates in different regions of the world. A default=0.089 Strawn2013aqraLm is taken. pSoil is conditional to several risk factors; yet this functions is based upon two risk factors:

irrigation or rain occurring prior to harvesting, whose odds-ratio has been estimated at 25.0 when taking place 24 hours before harvesting Weller2015qraLm; and

application of organic fertilizer, whose odds-ratio has been estimated at 7.0 when manure is applied within 1 year Strawn2013bqraLm.

The distribution about the concentration of L. monocytogenes in soil is represented as a triangular distribution, using data from Dowe1997;textualqraLm, who reported a mean of 4.0 MPN L. monocytogenes/ g soil, with a 95% confidence interval of \(<1.0-28\ MPN/g\).

References

statsqraLm

mc2dqraLm

Background2022qraLm

Dowe1997qraLm

Strawn2013aqraLm

Strawn2013bqraLm

Weller2015qraLm

Author

Laurent Guillier

Examples

# Considering all fields with foil: less cultivation lots contaminated than pSoil
caSoil2rind(nLots = 20, sizeLot = 10, pSoil = 0.089, pManure = 0.0, pIrrigRaining = 0, pFoil = 1.0)
#> $N
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    1    0    0    0    3    0    1     2
#>  [2,]    0    2    0    1    0    1    0    1    0     0
#>  [3,]    4    0    1    0    0    1    0    0    0     2
#>  [4,]    2    0    1    0    1    0    1    0    1     1
#>  [5,]    2    5    1    3    3    1    4    2    4     0
#>  [6,]    1    0    0    0    0    0    0    0    0     0
#>  [7,]    1    1    3    0    0    0    0    1    0     0
#>  [8,]    3    0    0    1    3    2    1    2    1     2
#>  [9,]    0    1    0    0    2    1    0    0    0     2
#> [10,]    0    1    0    0    0    0    2    0    0     0
#> [11,]    0    0    0    0    0    1    0    0    0     0
#> [12,]    1    0    0    0    1    0    0    1    1     0
#> [13,]    1    0    0    0    0    0    0    0    0     0
#> [14,]    0    0    2    0    1    0    0    1    0     0
#> [15,]    1    2    1    0    1    2    3    2    1     2
#> [16,]    1    0    3    0    0    1    0    1    0     0
#> [17,]    0    0    1    2    0    0    0    0    0     2
#> [18,]    0    1    2    2    4    1    0    0    0     5
#> [19,]    1    0    1    2    1    1    0    0    2     2
#> [20,]    1    0    0    1    3    0    0    2    0     1
#> 
#> $P
#> [1] 0.089
#> 

# Effect of manure (risk factor): more batches contaminated than pSoil
caSoil2rind(nLots = 20, sizeLot = 10, pSoil = 0.089, pManure = 0.5, pIrrigRaining = 0, pFoil = 0.0)
#> $N
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   22   39   51   48   16   34   23   18    6    38
#>  [2,]    0    3    1    0    1    1    0    0    0     0
#>  [3,]    1    0    1    0    2    1    1    1    0     1
#>  [4,]    0    1    1    3    1    2    4    2    1     1
#>  [5,]   14   12   11   25    6    2   17    6   24    19
#>  [6,]    1    9    3    2    4    0    8    0   17     3
#>  [7,]   20   20   16    2   14    5   12   19    1     4
#>  [8,]    1    1    0    1    1    0    2    0    3     1
#>  [9,]    3    0    1    5    3    1    3    7    7     1
#> [10,]    9    8    4    8   14    3    0    3    9    15
#> [11,]   17   12   26    7   18   18   33   24   19    31
#> [12,]   12    8    9    8   15    2   14    5    4     1
#> [13,]    2    2    2    0    4    1    2    3    1     1
#> [14,]    9    7    9    7   14   16    6    3   13    11
#> [15,]    1    6    1    0    9    9   13    8   13     6
#> [16,]   10   14   28   23    6    0    2   18   10     2
#> [17,]    0    2    2    0    3    0    1    2    4     2
#> [18,]   12   15    2    7   12    6    2    6   10     4
#> [19,]    2    2    1    7    9    0    1    4    2     7
#> [20,]    4    4    5    5    2    4    4    2    8     6
#> 
#> $P
#> [1] 0.1841383
#>