Route of contamination of cantaloupe from soil to rind

The function caSoil2rind() evaluates the contamination of cantaloupes through soil only. It considers soil contamination characteristics such as prevalence pSoil and concentration c_soil as well as the quantity of soil deposited on the cantaloupe rind. Furthermore, the algorithm allows for the use of foils to protect cantaloupe from soil contamination, and the user must provide the probability of growing cantaloupes using such protective foils. pSoil and c_soil have to be chosen by the user according to existing data; whereas pSoil is conditional to several risk factors. The identified risk factors (WHO 2022) are:

irrigation in the previous days before harvesting; and

the use of organic fertilizer.

which affect pSoil as associated odds-ratios (F_irrig_rain and fManure). In this context, the user has to define the proportion of fields practising irrigation prior to harvest or undergoing rain (p_irrig_rain), and the proportion of fields using organic amendments (pManure).

Usage

caSoil2rind(
  nLots,
  sizeLot,
  pSoil = 0.089,
  fManure = 7,
  pManure = 0.5,
  fIrrigRaining = 25,
  pIrrigRaining = 0.1,
  cSoilLogMin = -1,
  cSoilLogMode = 0.6,
  cSoilLogMax = 1.48,
  qSoilMin = 0.05,
  qSoilMode = 0.5,
  qSoilMax = 5,
  pFoil = 0.5,
  rFoil = 0.9
)

Arguments

nLots

Size of the Monte Carlo simulation (scalar).

sizeLot

Number of cantaloupes per cultivation lot or field (scalar).

pSoil

Probability of contamination of soil (provided by the user, \(default=0.089\) according to Strawn et al. (2013) )

fManure

Odds-ratio estimate associated to use or organic amendment in soil (\(default=7.0\) according to Strawn et al. (2013) )

pManure

Proportion of fields using organic amendments (provided by user, \(default=0.5\))

fIrrigRaining

Odds-ratio estimate associated to use of irrigation and raining events up to 2 days before harvest (\(default=25.0\) according to Weller et al. (2015) ).

pIrrigRaining

Proportion of fields that undergo irrigation or raining just previous harvest (provided by user, \(default=0.1\)).

cSoilLogMin

(log10 CFU/g) Minimum value of the triangular distribution describing variability of concentration (according to Dowe et al. (1997) \(default=-1\ log10\ CFU/g\)).

cSoilLogMode

(log10 CFU/g) Mode value of the triangular distribution describing variability of concentration (according to Dowe et al. (1997) \(default=0.6\ log10\ CFU/g\)).

cSoilLogMax

(log10 CFU/g) Maximum value of the triangular distribution describing variability of concentration (according to Dowe et al. (1997) : \(default=1.48\ log10\ CFU/g\)).

qSoilMin

(g) Minimum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\(default=0.05\ g\)).

qSoilMode

(g) Mode value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\(default=0.5\ g\)).

qSoilMax

(g) Maximum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\(default=5\ g\)).

pFoil

Proportion of fields grown in foil (e.g. plastic mulch) (EKE: \(default=0.5\))

rFoil

Reduction fraction of the quantity of soil transferred to rind (EKE: \(default=0.9\))

Value

A list of two elements:

N

(CFU) A matrix of size nLots by sizeLot containing the numbers of L. monocytogenes cells on cantaloupe from the soil route;

P

Prevalence of field lots of cantaloupes contaminated with L. monocytogenes from the soil route (scalar).

Note

The prevalence in soil, pSoil, must be provided by the user according to existing data. WHO (2022) lists various prevalence estimates in different regions of the world. A default=0.089 (Strawn et al. 2013) is taken. pSoil is conditional to several risk factors; yet this functions is based upon two risk factors:

irrigation or rain occurring prior to harvesting, whose odds-ratio has been estimated at 25.0 when taking place 24 hours before harvesting (Weller et al. 2015) ; and

application of organic fertilizer, whose odds-ratio has been estimated at 7.0 when manure is applied within 1 year (Strawn et al. 2013) .

The distribution about the concentration of L. monocytogenes in soil is represented as a triangular distribution, using data from Dowe et al. (1997) , who reported a mean of 4.0 MPN L. monocytogenes/ g soil, with a 95% confidence interval of \(<1.0-28\ MPN/g\).

References

Team RC (2022). R: A Language and Environment for Statistical Computing. R Foundation for Statistical Computing, Vienna, Austria. https://www.R-project.org/.

Pouillot R, Delignette-Muller M (2010). “Evaluating variability and uncertainty in microbial quantitative risk assessment using two R packages.” International Journal of Food Microbiology, 142(3), 330-40.

WHO (2022). “A Roadmap for the Development of Risk Assessment Models of Listeria monocytogenes in Selected Produce and Seafood Products.” World Health Organization.

Dowe MJ, Jackson ED, Mori JG, Bell CR (1997). “Listeria monocytogenes survival in soil and incidence in agricultural soils.” Journal of Food Protection, 60, 1201--1207.

Strawn LK, Fortes ED, Bihn EA, Nightingale KK, Gröhn YT, Worobo RW, Wiedmann M, Bergholz PW (2013). “Landscape and meteorological factors affecting prevalence of three food-borne pathogens in fruit and vegetable farms.” Applied and Environmental Microbiology, 79(2), 588-600. doi:10.1128/AEM.02491-12 .

Strawn LK, Gröhn YT, Warchocki S, Worobo RW, Bihn EA, Wiedmann M (2013). “Risk factors associated with Salmonella and Listeria monocytogenes contamination of produce fields.” Applied and Environmental Microbiology, 79(24), 7618-7627. doi:10.1128/AEM.02831-13 .

Weller D, Wiedmann M, Strawn LK (2015). “Spatial and temporal factors associated with an increased prevalence of Listeria monocytogenes in spinach fields in New York State.” Applied and Environmental Microbiology, 81(17), 6059-6069. doi:10.1128/AEM.01286-15 .

Author

Laurent Guillier

Examples

# Considering all fields with foil: less cultivation lots contaminated than pSoil
caSoil2rind(nLots = 20, sizeLot = 10, pSoil = 0.089, pManure = 0.0, pIrrigRaining = 0, pFoil = 1.0)
#> $N
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    1    0    0    0    3    0    1     2
#>  [2,]    0    2    0    1    0    1    0    1    0     0
#>  [3,]    4    0    1    0    0    1    0    0    0     2
#>  [4,]    2    0    1    0    1    0    1    0    1     1
#>  [5,]    2    5    1    3    3    1    4    2    4     0
#>  [6,]    1    0    0    0    0    0    0    0    0     0
#>  [7,]    1    1    3    0    0    0    0    1    0     0
#>  [8,]    3    0    0    1    3    2    1    2    1     2
#>  [9,]    0    1    0    0    2    1    0    0    0     2
#> [10,]    0    1    0    0    0    0    2    0    0     0
#> [11,]    0    0    0    0    0    1    0    0    0     0
#> [12,]    1    0    0    0    1    0    0    1    1     0
#> [13,]    1    0    0    0    0    0    0    0    0     0
#> [14,]    0    0    2    0    1    0    0    1    0     0
#> [15,]    1    2    1    0    1    2    3    2    1     2
#> [16,]    1    0    3    0    0    1    0    1    0     0
#> [17,]    0    0    1    2    0    0    0    0    0     2
#> [18,]    0    1    2    2    4    1    0    0    0     5
#> [19,]    1    0    1    2    1    1    0    0    2     2
#> [20,]    1    0    0    1    3    0    0    2    0     1
#> 
#> $P
#> [1] 0.089
#> 

# Effect of manure (risk factor): more batches contaminated than pSoil
caSoil2rind(nLots = 20, sizeLot = 10, pSoil = 0.089, pManure = 0.5, pIrrigRaining = 0, pFoil = 0.0)
#> $N
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   22   39   51   48   16   34   23   18    6    38
#>  [2,]    0    3    1    0    1    1    0    0    0     0
#>  [3,]    1    0    1    0    2    1    1    1    0     1
#>  [4,]    0    1    1    3    1    2    4    2    1     1
#>  [5,]   14   12   11   25    6    2   17    6   24    19
#>  [6,]    1    9    3    2    4    0    8    0   17     3
#>  [7,]   20   20   16    2   14    5   12   19    1     4
#>  [8,]    1    1    0    1    1    0    2    0    3     1
#>  [9,]    3    0    1    5    3    1    3    7    7     1
#> [10,]    9    8    4    8   14    3    0    3    9    15
#> [11,]   17   12   26    7   18   18   33   24   19    31
#> [12,]   12    8    9    8   15    2   14    5    4     1
#> [13,]    2    2    2    0    4    1    2    3    1     1
#> [14,]    9    7    9    7   14   16    6    3   13    11
#> [15,]    1    6    1    0    9    9   13    8   13     6
#> [16,]   10   14   28   23    6    0    2   18   10     2
#> [17,]    0    2    2    0    3    0    1    2    4     2
#> [18,]   12   15    2    7   12    6    2    6   10     4
#> [19,]    2    2    1    7    9    0    1    4    2     7
#> [20,]    4    4    5    5    2    4    4    2    8     6
#> 
#> $P
#> [1] 0.1841383
#>