Microbial reduction due to cooking at home

The fvCooking() function simulates the reduction in L. monocytogenes numbers in defrosted (or still frozen) vegetables due to cooking. The function can represent a scenario where the consumer gives the non-RTE frozen vegetables a non-intended use such as direct (uncooked) consumption in salads, smoothies, etc. In that case, the probability of cooking in the argument should be set to a value different from one. The variability in the effect of cooking defrosted (or frozen) vegetables is represented by a triangular distribution, as assumed in EFSA2020;textualqraLm.

Usage

fvCooking(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  pCooked = 0.825,
  minCook = 1,
  modeCook = 5,
  maxCook = 9
)

Arguments

data

a list of:

N

(CFU) A matrix containing the numbers of L. monocytogenes per portion of defrosted or frozen vegetables.

ProbUnitPos

Lot-specific probability of contaminated portions or servings, defrosted or not (vector).

P

Mean prevalence of contaminated portions of defrosted or frozen vegetables.

nLots

Number of lots sampled or size of the Monte Carlo simulation (scalar).

sizeLot

Number of units or portions produced in a lot (scalar).

pCooked

Probability of cooking the defrosted or frozen vegetables for consumption

minCook

(log10) Minimum value of the triangular distribution representing the variability in the reduction of L. monocytogenes in the event of cooking

modeCook

(log10) Mode value of the triangular distribution representing the variability in the reduction of L. monocytogenes in the event of cooking

maxCook

(log10) Maximum value of the triangular distribution representing the variability in the reduction of L. monocytogenes in the event of cooking

Value

A list of three elements:

N

(CFU) A matrix of size nLots lots by sizeLot units containing the numbers of L. monocytogenes in the portions or servings to be consumed;

ProbUnitPos

Lot-specific probability of contaminated portions to be consumed (vector);

P

Mean prevalence of contaminated portions or servings to be consumed (scalar).

Note

To represent variability in the effect of cooking defrosted (or frozen) vegetables, EFSA2020;textualqraLm assumed a triangular distribution, with parameters \(minCook=1\), \(modeCook=5\) and \(maxCook=9\). According to FSAI2022;textualqraLm, 84-89% of the respondants recognised that frozen vegetables need to be cooked prior to consumption, whereasWillis2020;textualqraLm found that 19% of the non-RTE frozen vegetables packages have no indication for cooking. An average value for the probability of cooking of \(p\_cooked=0.825\) can be used in this function.

References

mc2dqraLm

EFSA2020qraLm

Willis2020qraLm

FSAI2022qraLm

Author

Laurent Guillier

Examples

library(mc2d)
# N0 <- list(N=matrix(stats::rpois(2000, 15),
#            ncol=200, nrow=10),
#            P=0.12,
#            ProbUnitPos = 0.7)
#
# N_postcooking=fvCooking(N0,
#                          pCooked=0.825,
#                          minCook=1,
#                          modeCook=5,
#                          maxCook=9)
# hist(N_postcooking$N)