Cold chain logistics for ready-to-eat seafood

The function sfColdChain() simulates the growth of L. monocytogenes, as affected by lactic acid bacteria (LAB) populations, in RTE seafood during the various stages of the cold chain logistics, including transportation to retail, display at retail and transportation to home; and relies on the growth kinetic functions sfMejlholmDalgaard, sfMejlholmDalgaardLAB() and sfGrowthJameson(). The function sfColdChain() can be used for home storage as well, by changing only the parameters defining the time and temperature Pert distributions. For every lot, the algorithm samples time and temperature from Pert distributions defined by the parameters timeMin, timeMode, timeMax, and tempMin, tempMode, tempMax, respectively, provided by the user. The specific growth rates for L. monocytogenes and LAB are determined evaluating the intrinsic and extrinsic characteristics of the seafood product (sampled from the function sfCharacteristics()) in the Mejholm and Daalgard's existing secondary models for both bacterial groups ((Mejlholm and Dalgaard 2009) , (Mejlholm et al. 2010) , (Mejlholm and Dalgaard 2013) and (Mejlholm and Dalgaard 2015) ). The parameters of the two growth rate models related to reference growth rate, temperature, water activity, pH and concentrations of phenol, nitrites, CO2 at equilibrium, undissociated lactic acid, undissociated diacetate, undissociated acetic acid, undissociated benzoic acid, undissociated citric and undissociated sorbic acid, for L. monocytogenes and lactic acid bacteria, are set by default to those fitted and validated by Mejlholm et al. (2010) and Mejlholm and Dalgaard (2013) , respectively. However, any of the parameters of the growth rate models can be altered or updated. The numbers of L. monocytogenes in the RTE seafood product, as affected by LAB, are then estimated applying the extended Jameson-effect competition model published in Giménez and Dalgaard (2004) using the interaction gamma parameter proposed by Møller et al. (2013) .

Usage

sfColdChain(
  data = list(),
  RTE,
  nLots = NULL,
  sizeLot = NULL,
  unitSize = NULL,
  tempMin = 0.288,
  tempMode = 4.6,
  tempMax = 8.912,
  timeMin = 12,
  timeMode = 144,
  timeMax = 360,
  variability = "lot",
  corTimeTemp = 0,
  N0LABmin = 1.5,
  N0LABmode = 2.78,
  N0LABmax = 4.1,
  intralotSdN0LAB = 0,
  lnQ0LABmin = -12,
  lnQ0LABmode = -2.73,
  lnQ0LABmax = 1.26,
  MPDLABmin = 8,
  MPDLABmode = 8.5,
  MPDLABmax = 9,
  MPDLmmin = 6.6,
  MPDLmmode = 7.36,
  MPDLmmax = 8.2,
  mumaxrefLm = 0.419,
  TminLm = -2.83,
  TrefLm = 25,
  awminLm = 0.923,
  pHminLm = 4.97,
  pheMaxLm = 32,
  NITmaxLm = 350,
  CO2maxLm = 3140,
  micLACuLm = 3.79,
  micDACuLm = 4.8,
  micAACuLm = 10.3,
  micBACuLm = 0.349,
  micCACuLm = 2.119,
  micSACuLm = 1.896,
  mumaxrefLAB = 0.583,
  TminLAB = -5.25,
  TrefLAB = 25,
  awminLAB = 0.928,
  pHminLAB = 4.24,
  pheMaxLAB = 40.3,
  NITmaxLAB = 2780,
  CO2maxLAB = 6691,
  micLACuLAB = 12,
  micDACuLAB = 33.3,
  micAACuLAB = 10.3,
  micBACuLAB = 1.51,
  micCACuLAB = 10.3,
  micSACuLAB = 12.6,
  gamma = 1,
  lim = 1,
  step = 1
)

Arguments

data

a list of:

N

(CFU) A matrix containing the numbers of L. monocytogenes in packs of RTE seafood at the end of transport to retail, from contaminated lots.

P

Prevalence of contaminated lots (scalar).

N0LAB

(CFU) Optional: A matrix containing the numbers of LAB in packs of RTE seafood at the end of transport to retail, from contaminated lots. If missing, it will be generated from default parameters.

lnQ0Lm

Optional: A matrix containing the natural logarithm of q0 for L. monocytogenes in packs of RTE seafood. If missing, it will be generated from default parameters.

lnQ0LAB

Optional: A matrix containing the natural logarithm of q0 for LAB in packs of RTE seafood. If missing, it will be generated from default parameters.

MPDLm

(log10 CFU/g) Optional: A matrix containing the maximum population density of L. monocytogenes in packs of RTE seafood. If missing, it will be generated from default parameters.

MPDLAB

(log10 CFU/g) Optional: A matrix containing the log10 of MPD for LAB in packs of RTE seafood. If missing, it will be generated from default parameters.

RTE

Intrinsic and extrinsic characteristics of the RTE seafood product, provided using the sfCharacteristics function

nLots

Number of lots sampled or size of the Monte Carlo simulation (scalar).

sizeLot

Number of units or portions produced in a lot (scalar).

unitSize

(g) Weight of the contents of the RTE seafood product

tempMin

(\(^\circ C\)) Minimum storage temperature of the RTE seafood product (scalar). Will be used in a Pert distribution.

tempMode

(\(^\circ C\)) Mode of storage temperature of the RTE seafood product (scalar).

tempMax

(\(^\circ C\)) Maximum storage temperature of the RTE seafood product (scalar).

timeMin

(h) Minimum storage time of the RTE seafood product (scalar). Will be used in a Pert distribution.

timeMode

(h) Mode of storage time of the RTE seafood product (scalar).

timeMax

(h) Maximum storage time of the RTE seafood product (scalar).

variability

level of variability for time and temperature: "lot", "column" or "portion".

corTimeTemp

define

N0LABmin

(log10 CFU) Minimum value of the Pert distribution representing variability in the numbers of LAB at time=0 (scalar). Will be then used as a power of 10.

N0LABmode

(log10 CFU) Mode value of the Pert distribution representing variability in the numbers of LAB at time=0 (scalar).

N0LABmax

(log10 CFU) Maximum value of the Pert distribution representing variability in the numbers of LAB at time=0 (scalar).

intralotSdN0LAB

(log10 CFU) Intra-lot standard deviation to produce variability in the numbers of LAB (scalar).

lnQ0LABmin

Minimum value of the Pert distribution representing variability in the natural logarithm of the parameter q0 for LAB (scalar).

lnQ0LABmode

Mode value of the Pert distribution representing variability in the natural logarithm of the parameter q0 for LAB (scalar).

lnQ0LABmax

Maximum value of the Pert distribution representing variability in the natural logarithm of the parameter q0 for LAB (scalar).

MPDLABmin

(log10 CFU/g) Minimum value of the Pert distribution representing the variability in the maximum population density of LAB (scalar).

MPDLABmode

(log10 CFU/g) Mode value of the Pert distribution representing the variability in the maximum population density of LAB (scalar).

MPDLABmax

(log10 CFU/g) Maximum value of the Pert distribution representing the variability in the maximum population density of LAB (scalar).

MPDLmmin

(log10 CFU/g) Minimum value of the Pert distribution representing the variability in the maximum population density of L. monocytogenes (scalar).

MPDLmmode

(log10 CFU/g) Mode value of the Pert distribution representing the variability in the maximum population density of L. monocytogenes (scalar).

MPDLmmax

(log10 CFU/g) Maximum value of the Pert distribution representing the variability in the maximum population density of L. monocytogenes (scalar).

mumaxrefLm

(1/h) Maximum growth rate of L. monocytogenes at the reference temperature TrefLm (scalar).

TminLm

(\(^\circ C\)) Minimum temperature for growth of L. monocytogenes (scalar).

TrefLm

(\(^\circ C\)) Reference temperature for L. monocytogenes (scalar).

awminLm

Minimum water activity for growth of L. monocytogenes (scalar).

pHminLm

Minimum pH for growth of L. monocytogenes (scalar).

pheMaxLm

(ppm) Minimum inhibitory concentration of phenol for L. monocytogenes (scalar).

NITmaxLm

(ppm) Minimum inhibitory concentration of nitrites for L. monocytogenes (scalar).

CO2maxLm

(ppm) Minimum inhibitory concentration of CO2 for L. monocytogenes (scalar).

micLACuLm

(mM) Minimum inhibitory concentration of undissociated lactic acid for L. monocytogenes (scalar).

micDACuLm

(mM) Minimum inhibitory concentration of undissociated diacetate for L. monocytogenes (scalar).

micAACuLm

(mM) Minimum inhibitory concentration of undissociated acetic acid for L. monocytogenes (scalar).

micBACuLm

(mM) Minimum inhibitory concentration of undissociated benzoic acid for L. monocytogenes (scalar).

micCACuLm

(mM) Minimum inhibitory concentration of undissociated citric acid for L. monocytogenes (scalar).

micSACuLm

(mM) Minimum inhibitory concentration of undissociated sorbic acid for L. monocytogenes (scalar).

mumaxrefLAB

(1/h) Maximum growth rate of LAB at the reference temperature TrefLAB (scalar).

TminLAB

(\(^\circ C\)) Minimum temperature for growth of LAB (scalar).

TrefLAB

(\(^\circ C\)) Reference temperature for LAB (scalar).

awminLAB

Minimum water activity for growth of LAB (scalar).

pHminLAB

Minimum pH for growth of LAB (scalar).

pheMaxLAB

(ppm) Minimum inhibitory concentration of phenol for LAB (scalar).

NITmaxLAB

(ppm) Minimum inhibitory concentration of nitrites for LAB (scalar).

CO2maxLAB

(ppm) Minimum inhibitory concentration of CO2 for LAB (scalar).

micLACuLAB

(Mm) Minimum inhibitory concentration of undissociated lactic acid for LAB (scalar).

micDACuLAB

(Mm) Minimum inhibitory concentration of undissociated diacetate for LAB (scalar).

micAACuLAB

(Mm) Minimum inhibitory concentration of undissociated acetic acid for LAB (scalar).

micBACuLAB

(Mm) Minimum inhibitory concentration of undissociated benzoic acid for LAB (scalar).

micCACuLAB

(Mm) Minimum inhibitory concentration of undissociated citric acid for LAB (scalar).

micSACuLAB

(Mm) Minimum inhibitory concentration of undissociated sorbic acid for LAB (scalar).

gamma

Gamma parameter for the Jameson effect function. If 1, converges to classical Jameson effect (scalar).

lim

(log10 CFU/g) Limit to consider interactions in the Jameson effect (scalar).

step

(h) Integration step (scalar).

Value

A list of six elements:

N

(CFU) A matrix containing the numbers of L. monocytogenes in packs of RTE seafood after storage, from contaminated lots;

P

Prevalence of RTE seafood lots contaminated with L. monocytogenes (scalar);

N0LAB

(CFU) A matrix containing the numbers of LAB in packs of RTE seafood;

lnQ0Lm

A matrix containing the natural logarithm of q0 of L. monocytogenes in packs of RTE seafood;

lnQ0LAB

A matrix containing the natural logarithm of q0 of LAB in packs of RTE seafood;

SF

An object containing the intrinsic and extrinsic characteristics of the RTE seafood product.

Note

This function can be used multiple times. The microbial kinetic parameters MPD and ln_q0 for L. monocytogenes and LAB are sampled only once. Information for building the variability distributions about retail time and temperature was obtained from Pouillot et al. (2007) and FDA-FSIS (2003) . For Home storage, the Pert distribution for temperature can be defined using the following parameters: tempMin = qnorm(0.025, 7, 3), tempMode = 7, and tempMax = qnorm(0.975, 7, 3), which are reformulated from the normal(7.0, 3.0) used in Pouillot et al. (2007) . The Pert distribution for home storage time can be defined using the following parameters: timeMin = qweibull(0.025,shape = 1.14, scale = 18.39) * 24 timeMode = 18.39*((1.14-1)/1.14)^(1/1.14) * 24 timeMax = qweibull(0.975,shape = 1.14, scale = 18.39) * 24 which are reformulated from the Weibull(1.14, 18.39) days used in Endrikat et al. (2010) .

References

Endrikat S, Gallagher D, Pouillot R, Hicks Quesenberry H, Labarre D, Schroeder CM, Kause J (2010). “A comparative risk assessment for Listeria monocytogenes in prepackaged versus retail-sliced deli meat.” Journal of food protection, 73(4), 612-619. https://www.scopus.com/inward/record.uri?eid=2-s2.0-77951173159&doi=10.4315%2f0362-028x-73.4.612&partnerID=40&md5=8ca07e4a697bfdb1e5a2cf7944d5268e. Wolodzko T (2020). extraDistr: Additional Univariate and Multivariate Distributions. R package version 1.9.1, https://CRAN.R-project.org/package=extraDistr. FDA-FSIS (2003). “Interpretative Summary: Quantitative Assessment of the Relative Risk to Public Health from Foodborne Listeria monocytogenes among Selected Categories of Ready-to-Eat Foods.” Center for Food Safety and Applied Nutrition, Food and Drug Administration and Food Safety (US Department of Health and Human Services) and Inspection Service (US Department of Agriculture). Giménez B, Dalgaard P (2004). “Modelling and predicting the simultaneous growth of Listeria monocytogenes and spoilage micro-organisms in cold-smoked salmon.” Journal of Applied Microbiology, 96, 96-109. Pouillot R, Delignette-Muller M (2010). “Evaluating variability and uncertainty in microbial quantitative risk assessment using two R packages.” International Journal of Food Microbiology, 142(3), 330-40. Mejlholm O, Dalgaard P (2009). “Development and validation of an extensive growth and growth boundary model for Listeria monocytogenes in lightly preserved and ready-to-eat shrimp.” Journal of Food Protection, 72(10), 2132--2143. Mejlholm O, Gunvig A, Borggaard C, Blom-Hanssen J, Mellefont L, Ross T, Leroi F, Else T, Visser D, Dalgaard P (2010). “Predicting growth rates and growth boundary of Listeria monocytogenes—An international validation study with focus on processed and ready-to-eat meat and seafood.” International journal of food microbiology, 141(3), 137--150. Mejlholm O, Dalgaard P (2013). “Development and validation of an extensive growth and growth boundary model for psychrotolerant Lactobacillus spp. in seafood and meat products.” International Journal of Food Microbiology, 167(2), 244--260. Mejlholm O, Dalgaard P (2015). “Modelling and predicting the simultaneous growth of Listeria monocytogenes and psychrotolerant lactic acid bacteria in processed seafood and mayonnaise-based seafood salads.” Food Microbiology, 46, 1--14. Møller COA, Ilg Y, Aabo S, Christensen BB, Dalgaard P, Hansen TB (2013). “Effect of natural microbiota on growth of Salmonella spp. in fresh pork – A predictive microbiology approach.” Food Microbiology, 34(2), 284-295. ISSN 0740-0020, doi:10.1016/j.fm.2012.10.010 . Pouillot R, Miconnet N, Afchain A, Delignette-Muller ML, Beaufort A, Rosso L, Denis J, Cornu M (2007). “Quantitative risk assessment of Listeria monocytogenes in French cold-smoked salmon: I. Quantitative exposure assessment.” Risk Analysis, 27(3), 683-700. ISSN 0272-4332 0272-4332, doi:10.1111/j.1539-6924.2007.00921.x . Team RC (2022). R: A Language and Environment for Statistical Computing. R Foundation for Statistical Computing, Vienna, Austria. https://www.R-project.org/.

Author

Regis Pouillot rpouillot.work@gmail.com

Examples

sizeLot <- 100
size_mc <- 100
N <- matrix(round(10^rnorm(size_mc * sizeLot, 0, 3)),
            ncol = sizeLot,
            nrow = size_mc
            )

dat <- list(N = N, unitSize = 500)
RTE <- sfCharacteristics(size_mc)
str(RTE)
#> List of 12
#>  $ aw     : NULL
#>  $ NaCl   : num [1:100] 3.81 3.51 2.56 3.41 2.7 ...
#>  $ pH     : num [1:100] 6.14 6.25 6.14 6.04 5.96 ...
#>  $ P      : num [1:100] 7.11 12.37 9.95 16.38 9.88 ...
#>  $ CO2equi: num [1:100] 0.256 0.261 0.274 0.252 0.267 ...
#>  $ NIT    : num [1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ aaWph  : num [1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ baWph  : num [1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ caWph  : num [1:100] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ daWph  : num [1:100] 1552 1852 1298 1093 1734 ...
#>  $ laWph  : num [1:100] 17675 17920 13657 13565 13805 ...
#>  $ saWph  : num [1:100] 0 0 0 0 0 0 0 0 0 0 ...
dat1 <- sfColdChain(dat,
                    RTE = RTE
                    )
#> Integrate over  14  portions
#> ================================================================================
#> Warning: argument is not numeric or logical: returning NA
str(dat1)
#> List of 9
#>  $ N          : num [1:100, 1:100] 1290 0 0 16279 37 ...
#>  $ unitSize   : num 500
#>  $ lnQ0Lm     : num [1:10000] -8.04 -Inf -Inf -8.36 -8.33 ...
#>  $ MPDLm      : num [1:100] 7.44 7.84 7.22 7.11 7.35 ...
#>  $ MPDLAB     : num [1:100] 8.2 8.38 8.43 8.61 8.82 ...
#>  $ lnQ0LAB    : num [1:10000] 4.68 4.19 -4.5 -7.03 -1.82 ...
#>  $ N0LAB      : num [1:10000] 29948 41613 375 413 498 ...
#>  $ lotMeans   : num [1:100] 148.2 2084.7 33.6 5596.9 105.6 ...
#>  $ unitsCounts: num(0)