Growth of L. monocytogenes in raw fish

The function sfGrowthLPD() simulates the growth of L. monocytogenes in raw fish (salmon) considering lag phase duration. Parameters of the secondary models predicting specific growth rate and lag phase are taken from Jia et al. (2020) .

Usage

sfGrowthLPD(
  data = list(),
  unitSize = NULL,
  MPD = NULL,
  Temp,
  time,
  aLM = 0.0581,
  Tmin = 1.3,
  ALM = 0.84,
  mLM = 1.11,
  workDone = 0
)

Arguments

data

See Lot2LotGen() function.

unitSize

(g) Weight of a raw fish unit (whole or fillet).

MPD

(log10 CFU/g) Maximum population density of L. monocytogenes in raw fish, taken from Jia et al. (2020) .

Temp

(\(^\circ C\)) Temperature of raw fish fillet during holding time or storage.

time

(h) Length of holding time or storage.

aLM

Coefficient of the secondary model for growth rate, taken from Jia et al. (2020) .

Tmin

(\(^\circ C\)) Minimum temperature for growth of L. monocytogenes in raw fish, taken from Jia et al. (2020) .

ALM

Coefficient of the model for lag phase duration in hours, taken from Jia et al. (2020) .

mLM

Exponent of the model for lag phase duration in hours, taken from Jia et al. (2020) .

workDone

Work done during the stages prior to the current storage stage. Should be set to zero if there was no previous storage phase accounted for.

Value

A list of two elements:

N

(CFU) Numbers of L. monocytogenes in fish fillet after storage;

workDone

Work done during the previous stages including the current storage stage.

Note

The work-to-be-done concept from Daelman et al. (2013) was implemented to keep track of the lapse of the lag phase or the remaining lag phase during multiple storage stages.

References

Daelman J, Sharma A, Vermeulen A, Uyttendaele M, Devlieghere F, Membré J (2013). “Development of a time-to-detect growth model for heat-treated Bacillus cereus spores.” International Journal of Food Microbiology, 165(3), 231--240. ISSN 0168-1605, doi:/10.1016/j.ijfoodmicro.2013.04.018 .

Jia Z, Bai W, Li X, Fang T, Li C (2020). “Assessing the growth of Listeria monocytogenes in salmon with or without the competition of background microflora – A one-step kinetic analysis.” Food Control, 114. https://www.scopus.com/inward/record.uri?eid=2-s2.0-85082766072&doi=10.1016%2fj.foodcont.2020.107139&partnerID=40&md5=5d052a80577e8bb46aa9b7398c73767d.

Author

Regis Pouillot rpouillot.work@gmail.com and Ursula Gonzales-Barron ubarron@ipb.pt

Examples

dat <- Lot2LotGen(
  nLots = 100,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
N <- matrix(1:12, ncol = 4, nrow = 3)
dat <- list(N = N, unitSize = 500)

# Example of the first stage growth is taking place
aLM <- 0.0581 # parameter for mu (h-1)
Tmin <- 1.3 # parameter for mu (h-1)
ALM <- 0.84 # parameter for LPD (h)
mLM <- 1.11 # parameter for LPD (h)
MPD <- 9.2 # log10CFU/g # value taken from Jia et al. (2020)
time <- 11
Temp <- 2

first <- sfGrowthLPD(dat,
  MPD = MPD,
  Temp = 15,
  time = 11
)
#> Warning: argument is not numeric or logical: returning NA

# Example with previous stage where growth took place
second <- sfGrowthLPD(dat,
  MPD = MPD,
  Temp = 23,
  time = 11,
  workDone = first$workDone
)
#> Warning: argument is not numeric or logical: returning NA

# On matrix
third <- sfGrowthLPD(dat,
  MPD = MPD,
  Temp = c(10, 25, 30),
  time = 11
)
#> Warning: argument is not numeric or logical: returning NA