Secondary growth rate model of L. monocytogenes in RTE seafood

The function sfMejlholmDalgaard() estimates the growth rate of L. monocytogenes in RTE seafood, based on the predictive secondary model published by Mejlholm and Dalgaard (2009) , and later validated in RTE foods including seafood ((Mejlholm et al. 2010) ).

Usage

sfMejlholmDalgaard(
  Temp,
  aw = NULL,
  NaCl = 0,
  pH = 7,
  P = 0,
  CO2equi = 0,
  NIT = 0,
  aaWph = 0,
  baWph = 0,
  caWph = 0,
  daWph = 0,
  laWph = 0,
  saWph = 0,
  Tmin = -2.83,
  Tref = 25,
  awmin = 0.923,
  pHmin = 4.97,
  pheMax = 32,
  NITmax = 350,
  CO2max = 3140,
  micLACu = 3.79,
  micDACu = 4.8,
  micAACu = 10.3,
  micBACu = 0.349,
  micCACu = 2.119,
  micSACu = 1.896,
  mumaxref = 0.419
)

Arguments

Temp

(\(^\circ C\)) Storage temperature of the RTE product

aw

Water activity of the RTE product. If NULL, it will evaluate from NaCl

NaCl

(%) Salt content of the RTE product.

pH

pH of the RTE product.

P

(ppm) Phenol concentration in the RTE product.

CO2equi

(proportion) CO2 content in atmosphere in the RTE product (e.g. 0.25).

NIT

(ppm) Nitrites concentration in the RTE product.

aaWph

(ppm) Acetic acid concentration in the RTE product.

baWph

(ppm) Benzoic acid concentration in the RTE product.

caWph

(ppm) Citric acid concentration in the RTE product.

daWph

(ppm) Diacetate concentration in the RTE product.

laWph

(ppm) Lactic acid concentration in the RTE product.

saWph

(ppm) Sorbic acid concentration in the RTE product.

Tmin

(\(^\circ C\)) Minimum temperature for growth.

Tref

(\(^\circ C\)) Reference temperature.

awmin

Minimum water activity for growth.

pHmin

Minimum pH for growth.

pheMax

(ppm) MIC for phenols.

NITmax

(ppm) MIC for nitrites.

CO2max

(ppm) MIC for CO_2 .

micLACu

(mM) MIC for undissociated lactic acid.

micDACu

(mM) MIC for undissociated diacetate.

micAACu

(mM) MIC for undissociated acetic acid.

micBACu

(mM) MIC for undissociated benzoic acid.

micCACu

(mM) MIC for undissociated citric acid.

micSACu

(mM) MIC for undissociated sorbic acid.

mumaxref

(1/h) Maximum growth rate at the reference temperature Tref.

Value

mumax (1/h) Specific growth rate of L. monocytogenes in RTE seafood at the given intrinsic and extrinsic characteristics (scalar, vector or matrix).

Note

The parameters of the growth rate model are set by default to those fitted and validated by Mejlholm et al. (2010) . This model can also be found in the user friendly software FSSP: .

References

Mejlholm O, Dalgaard P (2009). “Development and validation of an extensive growth and growth boundary model for Listeria monocytogenes in lightly preserved and ready-to-eat shrimp.” Journal of Food Protection, 72(10), 2132--2143.

Mejlholm O, Gunvig A, Borggaard C, Blom-Hanssen J, Mellefont L, Ross T, Leroi F, Else T, Visser D, Dalgaard P (2010). “Predicting growth rates and growth boundary of Listeria monocytogenes—An international validation study with focus on processed and ready-to-eat meat and seafood.” International journal of food microbiology, 141(3), 137--150.

Author

Laurent Guillier

Examples

sfMejlholmDalgaard(Temp = c(-3, 15, 12), aw = 0.98, P = 3, saWph = 6000)
#> [1] 0.00000000 0.09583964 0.06630172