Secondary growth rate model of Lactic Acid Bacteria (LAB) in RTE seafood

The function sfMejlholmDalgaard() estimates the growth rate of LAB in RTE seafood, based on the predictive secondary model published by Mejlholm2013;textualqraLm, and later validated in RTE foods including seafood (Mejlholm2015qraLm).

Usage

sfMejlholmDalgaardLAB(
  Temp,
  aw = NULL,
  NaCl = 0,
  pH = 7,
  P = 0,
  CO2equi = 0,
  NIT = 0,
  aaWph = 0,
  baWph = 0,
  caWph = 0,
  daWph = 0,
  laWph = 0,
  saWph = 0,
  Tmin = -5.25,
  Tref = 25,
  awmin = 0.928,
  pHmin = 4.24,
  pheMax = 40.3,
  NITmax = 2780,
  CO2max = 6691,
  micLACu = 12,
  micDACu = 33.3,
  micAACu = 10.3,
  micBACu = 1.51,
  micCACu = 10.3,
  micSACu = 12.6,
  mumaxref = 0.583
)

Arguments

Temp

(\(^\circ C\)) Storage temperature of the RTE product

aw

Water activity of the RTE product. If NULL, it will evaluate from NaCl

NaCl

(%) Salt content of the RTE product.

pH

pH of the RTE product.

P

(ppm) Phenol concentration in the RTE product.

CO2equi

(proportion) CO2 content in atmosphere in the RTE product (e.g. 0.25).

NIT

(ppm) Nitrites concentration in the RTE product.

aaWph

(ppm) Acetic acid concentration in the RTE product.

baWph

(ppm) Benzoic acid concentration in the RTE product.

caWph

(ppm) Citric acid concentration in the RTE product.

daWph

(ppm) Diacetate concentration in the RTE product.

laWph

(ppm) Lactic acid concentration in the RTE product.

saWph

(ppm) Sorbic acid concentration in the RTE product.

Tmin

(\(^\circ C\)) Minimum temperature for growth.

Tref

(\(^\circ C\)) Reference temperature.

awmin

Minimum water activity for growth.

pHmin

Minimum pH for growth.

pheMax

(ppm) MIC for phenols.

NITmax

(ppm) MIC for nitrites.

CO2max

(ppm) MIC for CO_ 2.

micLACu

(mM) MIC for undissociated lactic acid.

micDACu

(mM) MIC for undissociated diacetate.

micAACu

(mM) MIC for undissociated acetic acid.

micBACu

(mM) MIC for undissociated benzoic acid.

micCACu

(mM) MIC for undissociated citric acid.

micSACu

(mM) MIC for undissociated sorbic acid.

mumaxref

(1/h) Maximum growth rate at the reference temperature Tref.

Value

mumax (1/h) Specific growth rate of LAB in RTE seafood at the given intrinsic and extrinsic characteristics (scalar, vector or matrix).

Note

The parameters of the growth rate model are set by default to those fitted and validated by Mejlholm2015;textualqraLm. This model can also be found in the user friendly software FSSP: .

References

Mejlholm2013qraLm

Mejlholm2015qraLm

Author

Laurent Guillier

Examples

# Reproducing screenshot value of FSSP (Fig. 1 red curve)
sfMejlholmDalgaardLAB(Temp = 5, aw = NULL, NaCl = 3.0, 
pH = 6.2, P = 5, laWph = 8000)
#> [1] 0.04265555
sfMejlholmDalgaardLAB(Temp = rnorm(10, 5, 0.2), 
                      aw = NULL, NaCl = 3.0, 
                      pH = 6.2, P = 5, laWph = 8000)
#>  [1] 0.04294988 0.04383470 0.04099391 0.04170544 0.03982608 0.04128490
#>  [7] 0.04011550 0.04199451 0.04128529 0.04443009